% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{add_terminal_label}
\alias{add_terminal_label}
\title{Add terminal label}
\usage{
add_terminal_label(
  data,
  entity_id_col = NULL,
  tax_level_labels = chemont_tax_levels
)
}
\arguments{
\item{data}{A \code{data.frame} of classified entities (or something that can be
coerced to one using \code{\link[base:as.data.frame]{base::as.data.frame()}}). Should have variables
corresponding to one or more of the items in \code{tax_level_labels}.}

\item{entity_id_col}{Character: a name, or vector of names, of variables in
\code{data} that uniquely identifies entities. Default \code{NULL} assumes that each
row is a unique entity.}

\item{tax_level_labels}{A character vector of all levels in the taxonomy.
Default is \code{\link{chemont_tax_levels}} to use the levels of the
ChemOnt taxonomy.}
}
\value{
The input \code{data.frame}, with new variables \code{terminal_label}
containing the terminal label for each entity, and \code{terminal_level}
containing the terminal taxonomy level for each entity (as an integer).
}
\description{
Add terminal classification labels
}
\details{
Adds the terminal classification label for each entity to a classified data
set.

The terminal classifications will be added in a new variable
\code{terminal_level}. If a variable with that name already exists in the data,
then it will be overwritten (with a warning).
}
\examples{
#add terminal labels to the first ten chemicals in the `biosolids_class` data set
 add_terminal_label(data = biosolids_class[1:10, ])
}
