% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{adjust_branch_lengths}
\alias{adjust_branch_lengths}
\title{Adjust branch lengths}
\usage{
adjust_branch_lengths(tree)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class).}
}
\value{
Numeric: A vector of adjusted branch lengths for the input tree.
}
\description{
Adjust branch lengths for pruned trees.
}
\examples{
my_tree <- prune_tree(tree = chemont_tree,
 prune_to = biosolids_class[1:10,])
adjust_branch_lengths(tree = tree)

}
\author{
Paul Kruse, Caroline Ring
}
