% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_comparison_dashboard.R
\name{batch_chemical_identifiers}
\alias{batch_chemical_identifiers}
\title{A function to collect chemical identifiers and store them in a data.table.}
\usage{
batch_chemical_identifiers(input_list)
}
\arguments{
\item{input_list}{A list of input values valid for \code{get_chemical_identifiers()}}
}
\value{
A data.table with the input list as a column, and chemical identifiers as subsequent columns
}
\description{
A function to collect chemical identifiers and store them in a data.table.
}
