% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BIOSOLIDS2021}
\alias{BIOSOLIDS2021}
\title{BIOSOLIDS2021 chemical list from the US EPA CompTox Chemicals Dashboard}
\format{
A csv representing 726 rows and 18 columns
\describe{
\item{DTXSID}{A chemical identifier}
\item{PREFERRED_NAME}{A character string of the name}
\item{CASRN}{A character string of the CASRN identifier}
\item{INCHIKEY}{A character of the InChIKey identifier}
\item{SMILES}{A character string of the SMILES identifier}
}
}
\usage{
BIOSOLIDS2021
}
\description{
A data set containing a list of 726 chemicals with related information
}
\details{
\url{https://comptox.epa.gov/dashboard/chemical-lists/BIOSOLIDS2021}
}
\seealso{
\code{\link{BIOSOLIDS2021_class}}
}
\keyword{datasets}
