% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BIOSOLIDS2021_class}
\alias{BIOSOLIDS2021_class}
\title{BIOSOLIDS2021 chemical list with ClassyFire classifications}
\format{
A csv representing 726 rows and 18 columns \describe{ \item{DTXSID}{A
chemical identifier} \item{PREFERRED_NAME}{A character string of the name}
\item{CASRN}{A character string of the CASRN identifier} \item{INCHIKEY}{A
character of the InChIKey identifier} \item{SMILES}{A character string of
the SMILES identifier} \item{kingdom}{ClassyFire kingdom label}
\item{superclass}{ClassyFire superclass label} \item{class}{ClassyFire
class label} \item{subclass}{ClassyFire subclass label} \item{level
5}{ClassyFire level 5 label} \item{level 6}{ClassyFire level 6 label}
\item{level 7}{ClassyFire level 7 label} \item{level 8}{ClassyFire level 8
label} \item{level 9}{ClassyFire level 9 label} \item{level 10}{ClassyFire
level 10 label} \item{level 11}{ClassyFire level 11 label} }
}
\usage{
BIOSOLIDS2021_class
}
\description{
A data set containing a list of 726 chemicals with related information
}
\details{
Original data source:
\url{https://comptox.epa.gov/dashboard/chemical-lists/BIOSOLIDS2021}.
ClassyFire classifications obtained via ClassyFire API.
}
\seealso{
\code{\link{BIOSOLIDS2021}}
}
\keyword{datasets}
