% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{calc_similarity}
\alias{calc_similarity}
\title{Calculate similarity measures}
\usage{
calc_similarity(tree, labels_A, labels_B, metric = "jaccard")
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{labels_A}{The first (set of) label(s).}

\item{labels_B}{The second (set of) label(s).}

\item{metric}{A string naming the similarity metric to use. Options are
"jaccard", "resnik", "lin", and "jiang_conrath". Only the first two letters
need be entered (e.g., "ja" for "jaccard", "re" for "resnik", "li" for
"lin", and "ji" for "jiang_conrath"). Case-insensitive.}
}
\value{
Matrix of similarity values for pairs of labels from \code{label_A} and
\code{label_B}.
}
\description{
Calculate similarity measures
}
\references{
\insertRef{pekar2002taxonomy}{treecompareR}

\insertRef{pesquita2009semantic}{treecompareR}
}
