% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{calc_similarity_data}
\alias{calc_similarity_data}
\title{Similarity for two datasets}
\usage{
calc_similarity_data(
  data_1,
  data_2,
  terminal_label = "terminal_label",
  tree = chemont_tree,
  tax_level_labels = chemont_tax_levels,
  entity_id_col = NULL,
  sim_matrix = NULL,
  similarity = NULL
)
}
\arguments{
\item{data_1}{A \code{data.frame} of classified entities}

\item{data_2}{Another \code{data.frame} of classified entities}

\item{terminal_label}{The variable name in the two data.frames that denotes
the terminal label of the classification. Default \code{"terminal_label"}.}

\item{tree}{The taxonomy tree to use as a \code{\link[ape]{phylo}}-class object (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class). Default \code{\link{chemont_tree}}.}

\item{tax_level_labels}{The set of taxonomy levels to use. Default
\code{\link{chemont_tax_levels}}.}

\item{entity_id_col}{\code{NULL} (default) or character vector giving the name(s)
of variables in \code{data} that specify unique entities. If \code{NULL}, each row
will be treated as a unique entity.}

\item{sim_matrix}{Optional: A pre-computed similarity matrix to use as a
lookup table. Default \code{NULL}, which will compute similarity from scratch
using the metric specified in \code{similarity}. If non-\code{NULL}, will override
anything specified in \code{similarity} (with a warning). For example, if \code{tree = chemont_tree}, you could use \code{sim_matrix = chemont_jaccard} to lookup
values in the pre-computed pairwise Jaccard similarity matrix for all nodes
in the ChemOnt taxonomy tree. Providing a pre-computed similarity matrix
may be faster if \code{data_1} and \code{data_2} are large.}

\item{similarity}{The similarity metric to calculate. Used only if
\code{sim_matrix = NULL}. Default \code{NULL}, which will calculate Jaccard
similarity. Options include "jaccard", "resnik", "lin", and
"jiang_conrath". Note: Ignored if \code{sim_matrix} is non-\code{NULL}, with a
warning.}
}
\value{
A similarity matrix with rows and columns corresponding to labels
from the \code{terminal_label} variable in \code{data_1} and \code{data_2}, respectively.
}
\description{
Calculate similarity measures for two datasets
}
\examples{
#computing similarity from scratch
calc_similarity_data(data_1 = biosolids_class[1:10,], data_2 = usgs_class[1:10, ], sim_matrix = NULL, similarity = "jaccard")

#using pre-computed similarity matrix as lookup table
calc_similarity_data(data_1 = biosolids_class[1:10,], data_2 = usgs_class[1:10, ], sim_matrix = chemont_jaccard)

#providing both sim_matrix and similarity throws a warning
calc_similarity_data(data_1 = biosolids_class[1:10, ], data_2 = usgs_class[1:20, ], sim_matrix = chemont_jaccard, similarity = "resnik")


}
\seealso{
\code{\link{jaccard_similarity}}, \code{\link{resnik_similarity}},
\code{\link{lin_similarity}}, \code{\link{jiang_conrath_similarity}},
\code{\link{similarity_matrix}}
}
\author{
Caroline Ring, Paul Kruse
}
