% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{check_resource}
\alias{check_resource}
\title{Check resource availability}
\usage{
check_resource(
  network_check_url = "http://httpstat.us/200",
  classyfire_check_url = "http://classyfire.wishartlab.com/"
)
}
\arguments{
\item{network_check_url}{Character: A URL to check that should never give an
HTTP error if the network is up. Default is
\code{"http://httpstat.us/200"}, which should always return status 200
(success). A \code{httr::HEAD()} request will be performed on this URL to check
whether the network is up.  If not, a message will be emitted and an empty
(placeholder) result will be returned.}

\item{classyfire_check_url}{Character: the URL to check to see whether the API is up.
Default is \code{"http://classyfire.wishartlab.com/"}. A \code{httr::HEAD()} request
will be performed on this URL to check whether the resource is up. If not, a
message will be emitted and an empty (placeholder) result will be returned.}
}
\value{
If all checks passed, returns \code{NULL}; otherwise, returns a string
containing a message about the URL checked and the error encountered.
}
\description{
Check to make sure ClassyFire API is available
}
\examples{
check_resource() #will return NULL if everything is OK


}
