% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{check_similarity_inputs}
\alias{check_similarity_inputs}
\title{Check similarity inputs}
\usage{
check_similarity_inputs(tree = NULL, node_1 = NULL, node_2 = NULL)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class)}

\item{node_1}{First node of interest. Character (a node name) or integer (a
node number).}

\item{node_2}{Second node of interest. Character (a node name) or integer (a
node number).}
}
\value{
Two-element numeric vector: A pair of node numbers corresponding to \code{node_1} and \code{node_2}.
}
\description{
Check to make sure nodes are valid.
}
\details{
This is a helper function for checking user input values within the
\code{\link{jaccard_similarity}}, \code{\link{resnik_similarity}},
\code{\link{lin_similarity}}, \code{\link{jiang_conrath_similarity}}, and
\code{link{similarity_matrix}} functions.
}
