% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctxR_methods.R
\name{chemical_identifiers}
\alias{chemical_identifiers}
\title{ctxR chemical identifiers}
\usage{
chemical_identifiers(input_list, verbose = FALSE)
}
\arguments{
\item{input_list}{A list or vector of CASRNS or DTXSIDS}

\item{verbose}{A Boolean indicating whether some 'progress report' is given.}
}
\value{
A data.table with columns \code{INPUT}, \code{DTXSID}, \code{CASRN},
\code{PREFFERED_NAME}, \code{SMILES}, \code{INCHIKEY} and each row corresponding to a
searched term.
}
\description{
ctxR chemical identifiers
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Grab identifiers
input_list <- c('DTXSID7020182', '67-64-1')
identifiers <- chemical_identifiers(input_list = input_list)
\dontshow{\}) # examplesIf}
}
