% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemical_list_biosolids_2022_05_10}
\alias{chemical_list_biosolids_2022_05_10}
\title{Sample US EPA CompTox Dashboard results for a data set}
\format{
A csv representing 726 rows and 18 columns
\describe{
\item{ï..DTXSID}
\item{PREFERRED_NAME}
\item{CASRN}
\item{INCHIKEY}
\item{IUPAC.NAME}
\item{SMILES}
\item{INCHI.STRING}
\item{MOLECULAR.FORMULA}
\item{AVERAGE.MASS}
\item{MONOISOTOPIC.MASS}
\item{X..of.SOURCES}
\item{X..OF.PUBMED.ARTICLES}
\item{PUBCHEM.DATA.SOURCES}
\item{CPDAT.COUNT}
\item{QC.Level}
\item{X..ToxCast.Active}
\item{X..ToxCast.Active.1}
\item{Total.Assays"}
}
}
\usage{
chemical_list_biosolids_2022_05_10
}
\description{
A data set containing a list of 726 chemicals with related information
}
\keyword{datasets}
