% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemont_df}
\alias{chemont_df}
\title{ChemOnt taxonomy as data.frame}
\format{
A data frame with 4825 rows and 3 columns:
\describe{
\item{Name}{ChemOnt text label for this node}
\item{ID}{ChemOnt ID number for this node}
\item{Parent_ID}{ChemOnt ID number for this node's parent}
}
}
\usage{
chemont_df
}
\description{
A data set describing the full ChemOnt taxonomy.
}
\details{
From \url{http://classyfire.wishartlab.com/tax_nodes.json}
}
\keyword{datasets}
