% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemont_jaccard}
\alias{chemont_jaccard}
\title{Jaccard similarity matrix for chemont tree}
\format{
A symmetric matrix with 4824 rows and 4824 columns:
\describe{
\item{rownames}{tip and node labels of chemont tree}
\item{colnames}{tip and node labels of chemont tree}
}
}
\usage{
chemont_jaccard
}
\description{
A matrix containing the Jaccard similarity values for all pairs of nodes in
the chemont tree.
}
\keyword{datasets}
