% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemont_lin_IC_SVH}
\alias{chemont_lin_IC_SVH}
\title{Lin similarity matrix for chemont tree}
\format{
A symmetric matrix with 4824 rows and 4824 columns:
\describe{
\item{rownames}{tip and node labels of chemont tree}
\item{colnames}{tip and node labels of chemont tree}
}
}
\usage{
chemont_lin_IC_SVH
}
\description{
A matrix containing the Lin similarity values for all pairs of nodes in
the chemont tree. This is derived using information content as defined in "An
intrinsic information content metric for semantic similarity in WordNet",
Seco et al 2004.
}
\keyword{datasets}
