% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemont_parent_child}
\alias{chemont_parent_child}
\title{Parent and child data for chemont tree}
\format{
A data frame with 4825 rows and 3 columns:
\describe{
\item{Name}{Name of the chemont label}
\item{ID}{ID of chemont label}
\item{Parent_ID}{ID of parent chemont label}
}
}
\usage{
chemont_parent_child
}
\description{
A data set containing the information required to build the chemont tree.
}
\keyword{datasets}
