% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemont_tree}
\alias{chemont_tree}
\title{Chemont tree}
\format{
A list of five objects:
\describe{
\item{edge}{A matrix with two columns and row for each edge in the tree.
The first column is the parent and the second column is the child.}
\item{Nnode}{The number of internal nodes, including the root.}
\item{tip.label}{A vector containing the tip labels corresponding to the
node numbers.}
\item{edge.length}{A vector containing edge lengths derived in the
construction of the \code{phylo} object. These are also useful for creating well
proportioned tree diagrams.}
\item{node.label}{A vector containing the node labels corresponding to the
node numbers.}
}
}
\usage{
chemont_tree
}
\description{
A \code{phylo} object encoding the ChemOnt ontology in its taxonomy form.
}
\keyword{datasets}
