% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{circ_tree_boxplot}
\alias{circ_tree_boxplot}
\title{Circular tree with boxplots}
\usage{
circ_tree_boxplot(
  data,
  col,
  tax_level_labels = chemont_tax_levels,
  title = NULL,
  tree = chemont_tree,
  layers = NULL,
  adjust_branch_length = FALSE,
  tippoint_boxplot = FALSE
)
}
\arguments{
\item{data}{A \code{data.frame} of classified entities with at least one
additional numeric data variable. Must contain at least the following
variables: variables named for every element of \code{tax_level_labels},
containing the (character) labels at each level of classification; and the
variable named in \code{col}.}

\item{col}{Character: the name of the numeric data variable to make boxplots
with.}

\item{tax_level_labels}{Character: the taxonomy levels. Default
\link{chemont_tax_levels}.}

\item{title}{Character: the title of the plot.}

\item{tree}{A \code{phylo}-class tree object specifying the tree to plot.}

\item{layers}{Which taxonomic layers to display outside of the boxplot layer.
This can be either a string with a single column name, a vector of column
names, or a list of column names. If the input is a vector or a list, it is
fine for it to be length 1.}

\item{adjust_branch_length}{\code{TRUE}/\code{FALSE}: If \code{TRUE}, adjust branch length
so that all newly-pruned terminal nodes appear at the same length as tips,
even if they were originally internal nodes. Default \code{FALSE}.}

\item{tippoint_boxplot}{\code{TRUE}/\code{FALSE}: Whether to color the tip points and
the boxplots. If \code{TRUE}, tip points and boxplots will have matching colors.
If \code{FALSE} (default), the boxplots will be filled white with black outline
and the tip points will be black.}
}
\value{
A \code{\link[ggtree:ggtree]{ggtree::ggtree()}} object, plotting the subtree induced by the data, annotated with boxplots
corresponding to the specified numeric column of data.
}
\description{
Plot a circular tree annotated with boxplots.
}
\details{
This function takes in a \code{data.frame} of classified entities with an
additional numeric data variable. The \code{data.frame} will be grouped
according to the terminal classification of the entities, and the specified
numeric data will be plotted as boxplot annotations at the corresponding
tips of the tree.
}
