% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{classify_by_smiles}
\alias{classify_by_smiles}
\title{SMILES string classification}
\usage{
classify_by_smiles(datatable)
}
\arguments{
\item{datatable}{A data.table that is the output from classify_datatable.}
}
\value{
A data.table object with classification information for each row.
}
\description{
This function takes a data.table of chemicals and classifications from
\code{\link{classify_datatable}} and attempts to classify chemicals from the
input data.table using SMILES strings that are missing classification data.
}
\seealso{
\code{\link{classify_datatable}}
}
