% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{cluster_tree}
\alias{cluster_tree}
\title{Heatmap cluster tree}
\usage{
cluster_tree(
  htmap,
  cluster_row,
  cluster_column,
  tree_obj,
  prune_to = NA,
  prune_args = list(keep_descendants = FALSE, adjust_branch_length = TRUE),
  ...
)
}
\arguments{
\item{htmap}{A \code{\link[ComplexHeatmap:HeatmapList]{ComplexHeatmap::HeatmapList()}} object produced using
\code{\link[=generate_heatmap]{generate_heatmap()}} with some row/column split specifications.}

\item{cluster_row}{Integer: the row index for the cluster of interest.}

\item{tree_obj}{The \code{phylo}-class tree object used to construct the heatmap
(i.e., the same \code{tree_obj} that was supplied to \code{\link[=generate_heatmap]{generate_heatmap()}}).}

\item{prune_to}{A pruning specification to apply to the tree visualization
(see \code{\link[=prune_tree]{prune_tree()}} for options). Default \code{NA}, which will prune the tree to
show only the union of the row and column labels in the heatmap. Use
\code{prune_to = NULL} (\emph{not} the default!) if you do not want to prune the base
tree.}

\item{prune_args}{A list of additional arguments to \code{\link[=prune_tree]{prune_tree()}}. Default
\code{list(keep_descendants = FALSE, adjust_branch_length = TRUE)}.}

\item{...}{Other arguments as for \code{\link[=display_subtree]{display_subtree()}}, excluding \code{base_tree}
\code{prune_to}, and \code{prune_args} (which are already supplied to this function as
\code{tree_obj}, \code{prune_to}, and \code{prune_args}), \code{highlight_by} (which will always
be \code{"set"} for this function), and \code{data_1}, \code{data_2}, \code{name_1}, and \code{name_2}
(which will be automatically determined from the heatmap row/column labels
and row/column titles).}

\item{cluster_column:}{Integer: the column index for the cluster of interest.}
}
\value{
A \code{\link[ggtree:ggtree]{ggtree::ggtree()}} object produced by \code{\link[=display_subtree]{display_subtree()}}.
}
\description{
Tree visualization showing membership in a heatmap cluster
}
\details{
Given a heatmap produced using \code{\link[=generate_heatmap]{generate_heatmap()}}, with clusters
produced using some row/column splitting specifications (\code{row_split} and
\code{column_split} arguments to \code{\link[=generate_heatmap]{generate_heatmap()}}), produce a tree diagram
with branches highlighted by their membership in the heatmap row labels for
that cluster, and the heatmap column labels for that cluster.
}
\examples{
#generate a heatmap with clusters
my_htmap <- generate_heatmap(
tree_object = chemont_tree,
matrix = chemont_jaccard,
row_data = biosolids_class,
column_data = usgs_class,
terminal_only = TRUE,
row_split = 6L,
column_split = 6L,
row_title = "Biosolids",
column_title = "USGS Water",
row_gap = unit(0.5, "mm"),
column_gap = unit(0.5, "mm"))

#now show a tree higlighted by membership in cluster 1/1 (high similarity)

cluster_tree(htmap = my_htmap,
cluster_row = 1,
cluster_column = 1,
tree_obj = chemont_tree,
show_tiplabs = FALSE,
clade_level = 2,
bg_tree_scale = 0,
base_opts = list(size = 1))

#compare to a tree highlighted by row/column labels in cluster 2/3 (some high and some low similarity)

#'cluster_tree(htmap = my_htmap,
cluster_row = 2,
cluster_column = 3,
tree_obj = chemont_tree
show_tiplabs = FALSE,
clade_level = 2,
bg_tree_scale = 0,
base_opts = list(size = 1))

}
