% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_comparison_dashboard.R
\name{convert_smiles_to_hex}
\alias{convert_smiles_to_hex}
\title{Helper function for converting SMILES string to API acceptable format}
\usage{
convert_smiles_to_hex(smiles = NULL)
}
\arguments{
\item{smiles}{A string representing a chemical in SMILES format}
}
\value{
A string converted for use in API calls.
}
\description{
Helper function for converting SMILES string to API acceptable format
}
