% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{count_descendants}
\alias{count_descendants}
\title{Count descendants}
\usage{
count_descendants(tree)
}
\arguments{
\item{tree}{An object of class \code{phylo} (see \code{\link[ape:read.tree]{ape::read.tree()}} for
description of this class)}
}
\value{
A \code{data.frame} with variables \code{node} (the node number), \code{descendants}
(the number of descendants for that node), \code{children} (the number of direct
children for that node), and \code{level}.
}
\description{
Count the number of descendants for each tree node
}
\details{
Count the number of descendants and direct children for each
node of the input tree.
}
\examples{
  #a randomly-generated tree
  tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)
  #count descendants of each node in this tree
  count_descendants(tree = tree)

}
