% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{count_entities_per_label}
\alias{count_entities_per_label}
\title{Count entities per label}
\usage{
count_entities_per_label(
  data,
  entity_id_col = NULL,
  tax_level_labels = chemont_tax_levels
)
}
\arguments{
\item{data}{A \code{data.frame} of classified entities.}

\item{entity_id_col}{\code{NULL} (default) or character vector giving the name(s)
of variables in \code{data} that specify unique entities. If \code{NULL}, each row
will be treated as a unique entity.}

\item{tax_level_labels}{A character vector of taxonomy levels. Default is
\code{\link{chemont_tax_levels}}, the levels of the ClassyFire taxonomy.}
}
\value{
A list the same length as , and named after, \code{tax_level_labels},
where each element is a \code{data.frame} with first variable \code{label}, giving
the labels, and second variable \code{n}, giving the count of entities with each
label in the data.
}
\description{
Given a \code{data.frame} of classified entities, determine the number of entities
with each label at each level of the taxonomy.
}
\examples{
#count entities per label in the first ten BIOSOLIDS2021 chemicals
count_entities_per_label(data = biosolids_class[1:10, ])

}
