% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{count_labels}
\alias{count_labels}
\title{Count unique labels}
\usage{
count_labels(data, tax_level_labels = chemont_tax_levels)
}
\arguments{
\item{data}{A \code{data.frame} of classified entities.}

\item{tax_level_labels}{A character vector of taxonomy levels. Default is
\code{\link{chemont_tax_levels}}, the levels of the ClassyFire taxonomy.}
}
\value{
The number of labels per taxonomy level.
}
\description{
Count the number of unique labels per taxonomy level in a classified data set.
}
\examples{
#count unique labels in the first ten BIOSOLIDS2021 chemicals
count_labels(data = biosolids_class[1:10, ])

}
