% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{data_set_subtrees}
\alias{data_set_subtrees}
\title{This function creates two plots that show a visual comparison of the subtrees induced by two data sets}
\usage{
data_set_subtrees(
  data_1,
  data_2,
  name_1 = "data_1",
  name_2 = "data_2",
  tax_level_labels = NULL,
  tree = NULL,
  show_tips = TRUE
)
}
\arguments{
\item{data_1}{A data.table of chemicals and their classifications.}

\item{data_2}{A data.table of chemicals and their classifications.}

\item{name_1}{An alternate parameter giving the name of the first data set.}

\item{name_2}{An alternate parameter giving the name of the second data set.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}

\item{tree}{An alternate parameter giving a taxonomy if not using ChemOnt.}

\item{show_tips}{An alternate parameter determining whether to show tip labels.}
}
\value{
A list of two ggtree objects.
}
\description{
This function creates two plots that show a visual comparison of the subtrees induced by two data sets
}
