% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{drop_tips_nodes}
\alias{drop_tips_nodes}
\title{Drop tips and nodes}
\usage{
drop_tips_nodes(
  tree,
  data = NULL,
  labels = NULL,
  nodes = NULL,
  level = NULL,
  keep_descendants = NULL,
  tax_level_labels = chemont_tax_levels,
  entity_id_col = NULL
)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class).}

\item{data}{A \code{data.frame} of classified entities. Default \code{NULL}.}

\item{labels}{Character: a set of labels of the subtree. Default \code{NULL}.}

\item{nodes}{Integer: a set of nodes of the subtree. Default \code{NULL}.}

\item{level}{Integer: the level to which the tree should be pruned. The root
is level zero and each subsequent generation of children nodes is one level
greater. Default \code{NULL}.}

\item{keep_descendants}{\code{TRUE}/\code{FALSE}: whether to keep all descendants of
the input labels/nodes/level. Default \code{NULL}, which means behavior depends
on whether \code{data}, \code{labels}, \code{nodes}, or \code{level} was provided. See
\code{\link[=prune_tree]{prune_tree()}} for more details.}

\item{tax_level_labels}{Character vector naming the levels of the taxonomy in
order from most general to most specific (excluding root). Default
\code{\link[=chemont_tax_levels]{chemont_tax_levels()}}.}

\item{entity_id_col}{\code{NULL} (default) or character vector giving the name(s)
of variables in \code{data} that specify unique entities. If \code{NULL}, each row
will be treated as a unique entity.}
}
\value{
A \code{phylo}-class object representing the induced subtree of the data.
}
\description{
Drop tips and nodes of a tree as needed to keep only the nodes
of a specified subtree.
}
\details{
This is a helper function used by \code{\link[=prune_tree]{prune_tree()}}. Usually it will
not be called by the user directly.

This function takes in a \code{phylo}-class tree and one of several ways of
specifying a subtree, and drops tips and nodes until the remaining tree
consists solely of the nodes and tips associated with the specified subtree
and their ancestors. The subtree can be specified using argument \code{data} as
a \code{data.frame} of classified entities; using argument \code{label} as a
character vector of subtree labels; using argument \code{nodes} as an integer
vector of tip and node IDs; or using argument \code{level} as either a character
string naming one of the taxonomy levels, or an integer giving the taxonomy
level of interest (where root is level 0).
}
\examples{
drop_tips_nodes(tree = chemont_tree,
data = biosolids_class[1:20,])


}
\references{
\insertRef{apepackage}{treecompareR}
}
\author{
Caroline Ring, Paul Kruse
}
