% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{enumerate_partitions}
\alias{enumerate_partitions}
\title{Enumerate partitions}
\usage{
enumerate_partitions(n, max_deg = NULL, min_deg = NULL)
}
\arguments{
\item{n}{An integer for which partitions will be generated and enumerated.}

\item{max_deg}{Integer: maximum permissible
partitions. Default \code{NULL} to auto-select.}

\item{min_deg}{Integer: minimum permissible
partitions. Default \code{NULL} to auto-select.}
}
\value{
A named list of partitions, with names corresponding to probabilities.
}
\description{
Enumerate partitions
}
\examples{
enumerate_partitions(n = 5)


}
\author{
Paul Kruse
}
