% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{general_Jaccard_similarity}
\alias{general_Jaccard_similarity}
\title{This determines the Jaccard similarity for two input node labels in a given tree.}
\usage{
general_Jaccard_similarity(tree, label_A, label_B)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{label_A}{The first label.}

\item{label_B}{The second label.}
}
\value{
The Jaccard distance of the label sets for the root to node path.
}
\description{
This determines the Jaccard similarity for two input node labels in a given tree.
}
