% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{general_Lin_similarity}
\alias{general_Lin_similarity}
\title{This determines the Lin similarity for two input nodes in a given tree.}
\usage{
general_Lin_similarity(
  tree,
  label_A = NULL,
  label_B = NULL,
  node_A = NULL,
  node_B = NULL
)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree, with an information content attribute IC.}

\item{label_A}{The first node label.}

\item{label_B}{The second node label.}

\item{node_A}{Alternate parameter, the first node number.}

\item{node_B}{Alternate parameter, the second node number.}
}
\value{
The Lin similarity in the given tree of the pair of nodes.
}
\description{
This determines the Lin similarity for two input nodes in a given tree.
}
