% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{general_Resnik_similarity}
\alias{general_Resnik_similarity}
\title{Resnik similarity}
\usage{
general_Resnik_similarity(
  tree,
  label_A = NULL,
  label_B = NULL,
  node_A = NULL,
  node_B = NULL
)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree, with an information
content attribute IC.}

\item{label_A}{The first node label.}

\item{label_B}{The second node label.}

\item{node_A}{Alternate parameter, the first node number.}

\item{node_B}{Alternate parameter, the second node number.}
}
\value{
The Resnik similarity in the given tree of the pair of nodes.
}
\description{
This determines the Resnik similarity for two input nodes in a given tree.
This uses the formulation as described in
\href{https://www.researchgate.net/publication/220837848_An_Intrinsic_Information_Content_Metric_for_Semantic_Similarity_in_WordNet/stats}{An
Intrinsic Information Content Metric for Semantic Similarity in WordNet}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)
\dontshow{\}) # examplesIf}
}
\references{
\insertRef{lin1998information}{treecompareR}

\insertRef{resnik1995using}{treecompareR}
}
