% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{generate_information_content}
\alias{generate_information_content}
\title{Generate information content.}
\usage{
generate_information_content(tree)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class)}
}
\value{
Numeric: a vector of information content for each node of the tree.
}
\description{
Generate information content for a tree.
}
\details{
Generates a data.frame of information content for the input
tree, where information content is as described in
\href{https://www.researchgate.net/publication/220837848_An_Intrinsic_Information_Content_Metric_for_Semantic_Similarity_in_WordNet/stats}{An
Intrinsic Information Content Metric for Semantic Similarity in WordNet}. The
data.frame also includes the depth of each tip and internal node, the number
of descendants of each node, and the number of children for each node.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)

generate_information_content(tree = tree)
\dontshow{\}) # examplesIf}
}
\references{
\insertRef{seco2004intrinsic}{treecompareR}
}
