% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_tree.R
\name{generate_parent_child}
\alias{generate_parent_child}
\title{Generate parent child from table}
\usage{
generate_parent_child(data, col_indices, id_index, root_label = "Root_label")
}
\arguments{
\item{data}{A data.frame with columns for each taxonomic level, and a row for
each node/tip, organized in a tree structure (children rows immediately
follow their parent rows)}

\item{col_indices}{Columns of taxonomic levels}

\item{id_index}{A column giving taxonomic id label}

\item{root_label}{The root label}
}
\value{
A data.frame with the parent child relationships for each node/tip in
the tree. These are encoded in three columns 'Name', 'ID', 'Parent_ID' and
each row corresponds to a unique node/tip.
}
\description{
Generate parent child from table
}
