% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_partition}
\alias{generate_partition}
\title{Partition generator}
\usage{
generate_partition(n)
}
\arguments{
\item{n}{positive integer at least 2.}
}
\value{
A vector of integers each at least 2 and with sum equal to the input.
}
\description{
Randomly generate partition
}
\details{
This function generates a partition of an input positive integer. Each
constituent of the partition has value at least 2.
}
\examples{
set.seed(42)
generate_partition(n = 4)

}
\author{
Paul Kruse
}
