% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{generate_similarity_matrix}
\alias{generate_similarity_matrix}
\title{Similarity matrix generator}
\usage{
generate_similarity_matrix(tree, similarity = NULL)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{similarity}{A similarity measure function that requires as inputs a
tree, label_A, and label_B}
}
\value{
A similarity matrix, which is a symmetric, with values in
\eqn{[0,1]}.
}
\description{
This function takes in a tree and a similarity measure function, and
generates a similarity matrix for the given tree and similarity measure. The
matrix is symmetric, with rows and columns given by the tip and internal node
labels in the ordering given by the tree.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)
\dontshow{\}) # examplesIf}
}
