% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_star}
\alias{generate_star}
\title{Generate Star tree}
\usage{
generate_star(n)
}
\arguments{
\item{n}{The number of tips.}
}
\value{
A \code{phylo} object representing the generated tree. (See
\code{\link[ape:read.tree]{ape::read.tree()}} for description of the \code{phylo} class.)
}
\description{
This function generates a star tree, a rooted tree of minimal
depth with n tips.
}
\examples{
set.seed(42)
generate_star(5)
}
\author{
Paul Kruse
}
