% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_tree.R
\name{generate_tree}
\alias{generate_tree}
\title{Convert ontology from data.frame to 'phylo' object.}
\usage{
generate_tree(file_dir, dataframe = NULL)
}
\arguments{
\item{file_dir}{A path to a JSON file}

\item{dataframe}{An optional data.frame containing parent_child relationship}
}
\value{
A list containing a 'phylo' object and data.frame with useful
information for building the phylo object.
}
\description{
Convert ontology from data.frame to 'phylo' object.
}
