% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_all_clades}
\alias{get_all_clades}
\title{List all clades in a tree at a specified level}
\usage{
get_all_clades(tree, level)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class).}

\item{level}{The level at which to display nodes (0 is the root)}
}
\value{
A data.frame with four variables: \code{node} (the node number in the
tree); \code{level} (the level of the node in the tree, where root is level
0); \code{parent} (the node number of the node's immediate parent); and
\code{Name} (the text label of the node).
}
\description{
List all clades in a tree at a specified level
}
\examples{
#list all ChemOnt superclasses:
get_all_clades(tree = chemont_tree, level = 2)

}
\author{
Caroline Ring
}
