% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{get_label_level}
\alias{get_label_level}
\title{Get labels at a specified taxonomy level}
\usage{
get_label_level(data, level_label, tax_level_labels = chemont_tax_levels)
}
\arguments{
\item{data}{A \code{data.frame} of classified entities (or something that can be
coerced to one using \code{\link[base:as.data.frame]{base::as.data.frame()}})}

\item{level_label}{Character: A taxonomy level of the classified data.}

\item{tax_level_labels}{A character vector of all of the taxonomy levels, in order.
Default is \code{\link{chemont_tax_levels}}, the levels of the ChemOnt
taxonomy.}
}
\value{
Character vector: The unique labels in \code{data} at the given taxonomy
level.
}
\description{
Get all the unique labels for a given taxonomy level in a given set of
classified entities.
}
\examples{
#get all the superclass labels in the BIOSOLIDS2021 data
get_label_level(data = biosolids_class,
level_label = "superclass")

}
