% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{get_labels}
\alias{get_labels}
\title{Get labels}
\usage{
get_labels(data, tax_level_labels = chemont_tax_levels)
}
\arguments{
\item{data}{A \code{data.frame} of classified entities (or something that can be
coerced to one using \code{\link[base:as.data.frame]{base::as.data.frame()}})}

\item{tax_level_labels}{A character vector of all of the taxonomy levels, in
order. Default is \code{\link{chemont_tax_levels}}, the levels of the
ChemOnt taxonomy.}
}
\value{
A \code{list} the same length as \code{tax_level_labels}: character vectors
containing the unique classification labels in the data for each level of
taxonomy.
}
\description{
Retrieve all labels in a \code{data.frame} of classified entities, grouped by
taxonomy level.
}
\examples{
#get labels for the first ten chemicals in the classified biosolids set
get_labels(data = biosolids_class[1:10, ])

}
