% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_node_level}
\alias{get_node_level}
\title{Node level}
\usage{
get_node_level(tree, node)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class).}

\item{node}{A vector: Character (node names) or integer (node ID numbers).}
}
\value{
An integer vector the same length as \code{node}, giving the taxonomic
level of the node(s) (where level 0 is the root).
}
\description{
Get the level of the given node(s) in a rooted tree.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

  tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)

  get_node_level(tree = tree, node = 't2') #specify node by name
  get_node_level(tree = tree, node = 2) #specify node by number
\dontshow{\}) # examplesIf}
}
