% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{get_number_of_labels}
\alias{get_number_of_labels}
\title{Number of labels}
\usage{
get_number_of_labels(data, tax_level_labels = chemont_tax_levels)
}
\arguments{
\item{data}{A data.table of chemicals with classifications.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not  using ClassyFire taxonomy}
}
\value{
A named list of chemical labels and their number of occurrences.
}
\description{
This is a helper function to determine number of occurrences for each label
in a data.table containing chemicals and their classification data.
}
