% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{get_overlap}
\alias{get_overlap}
\title{Get overlap}
\usage{
get_overlap(grouplab, group_col, entity_id_col = NULL, data_1, data_2)
}
\arguments{
\item{grouplab}{Character: the group label}

\item{group_col}{Character: the variable name for the group label (i.e., the
taxonomy level)}

\item{entity_id_col}{Character vector: variable name(s) in \code{data_1} and
\code{data_2} that define the unique entity ID}

\item{data_1}{A data.frame of classified entities}

\item{data_2}{A data.frame of classified entities}
}
\value{
A \code{data.frame} with variables \code{group} (containing \code{grouplab}),
\code{n_intersect} (containing the number of intersecting entities in that group
for the two datasets), \code{n_union} (containing the number of entities in the
union of that group for the two datasets), and \code{simil} (containing the
fractional similarity or overlap of entities in that group between the two
datasets)
}
\description{
Helper function for \code{\link[=calculate_overlap]{calculate_overlap()}}
}
\details{
Calculates number of overlapping entities in two data sets for a particular
classification label.
}
\examples{
get_overlap(grouplab = "Organic compounds",
group_col = "kingdom",
data_1 = biosolids_class,
data_2 = usgs_class)

}
