% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_tree.R
\name{get_parent_child}
\alias{get_parent_child}
\title{Reads in JSON file with parent-child relationship for taxonomy}
\usage{
get_parent_child(file_name)
}
\arguments{
\item{file_name}{Path to JSON file}
}
\value{
A data.frame with parent-object relationships.
}
\description{
Reads in JSON file with parent-child relationship for taxonomy
}
