% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_tree.R
\name{get_path_to_root}
\alias{get_path_to_root}
\title{Path to root}
\usage{
get_path_to_root(df, x)
}
\arguments{
\item{df}{data.frame with parent-child relationships}

\item{x}{Numeric identification number for node}
}
\value{
Path from node to root, as a string delimited by '/'.
}
\description{
This helper function generates a path from the root to a given node from a
data.frame of parent-child relationships. This is used in creating a tree
based off of the data.frame of parent-child relationships.
}
