% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{get results}
\usage{
get_results(
  url,
  label = "query",
  retry_get_times = 3,
  wait_sec = 5,
  terminate_on = c(400:407, 409:418, 421:428, 431, 451, 500:511),
  type = "structure"
)
}
\arguments{
\item{url}{As for \code{\link[=query_classyfire]{query_classyfire()}}. Character: A URL for a GET query.
Usually, this should be of format
\code{'http://classyfire.wishartlab.com/queries/NNNNN.json'}, where NNNNN are one
or more digits 0-9 denoting a query ID number.}

\item{label}{As for \code{\link[=query_classyfire]{query_classyfire()}}. Character: A label for the query.
Default \code{"query"}.}

\item{retry_get_times}{As for \code{\link[=query_classyfire]{query_classyfire()}}. Integer: number of times
to retry GET query if HTTP error was returned (except for errors in
\code{terminate_on}, which will terminate immediately without retrying).}

\item{wait_sec}{As for \code{\link[=query_classyfire]{query_classyfire()}}. Integer: number of seconds to
wait between retries. Minimum 5.}

\item{terminate_on}{Integer vector: List of HTTP status codes which will
immediately terminate with no more retries. Default: \code{ c(400:407, 409:418, 421:428, 431, 451, 500:511)}}
}
\value{
A list containing the parsed JSON results (if query was successful),
or placeholder values for the expected elements of the parsed JSON results.
Will contain the following named elements: \code{id}, \code{query_url}, \code{query_status}, \code{label},
\code{classification_status}, \code{number_of_elements}, \code{number_of_pages},
\code{invalid_entities}, and \code{entities}. See Details.
}
\description{
Helper function to get ClassyFire query results
}
\details{
This is a helper function used by \code{\link[=query_classyfire]{query_classyfire()}}. It generally
should not be called directly by the user.
}
\section{Return value details}{
\itemize{
\item \code{id}: Character: the string giving the ClassyFire query id. This is parsed from function argument \code{url}. If \code{url} is not of the expected structure, \code{id} will be \code{NA_character_}.
\item \code{url}: Character: The URL of the request, as provided in function argument \code{url}.
\item \code{query_status}: Character: The HTTP status message of the GET request (if one was available), or the error message if no HTTP status was available.
\item \code{label}: Character: The text label assigned to the query, if any.
\item \code{classification_status}: Character: the status of the ClassyFire classification. One of \code{'Done'}, \code{'In Queue'}, \code{'Processing'}, or \code{'Failed'}.
\item \code{number_of_elements}: Integer: the number of elements returned. If something went wrong, the placeholder value is \code{NA_integer_}.
\item \code{number_of_pages}: Integer: the number of pages of results. If something went wrong, the placeholder value is 1.
\item \code{invalid_entities}: A \code{data.frame} of entities in the query whose structures were invalid and no classification could be made. May be an empty \code{list} if there were no invalid entities, or if classification failed. If not empty, will be a \code{data.frame} with variables \code{structure} (character, containing the input structure) and \code{report} (character, explaining what went wrong).
\item \code{entities}: A nested \code{data.frame} with information about entities in the query whose structures were valid and which could be classified.
}

\code{entities} should be a \code{data.frame} with the following variables:
\itemize{
\item \code{identifier} Character: the entity identifier.
\item \code{smiles} Character: the entity SMILES string.
\item \code{inchikey} Character: the entity InChIKey.
\item \code{kingdom} A \code{data.frame} with the entity's kingdom-level classification.
\item \code{superclass} A \code{data.frame} with the entity's superclass-level classification.
\item \code{class} A \code{data.frame} with the entity's class-level classification.
\item \code{subclass} A \code{data.frame} with the entity's subclass-level classification.
\item \code{intermediate_nodes} A \code{list} of \code{data.frame}s, one list element for each entity. Each \code{data.frame} will contain information about any intermediate-level classifications between subclass-level and direct parent.
\item \code{direct_parent} A \code{data.frame} with the entity's direct parent classification (i.e. the most specific level of classification available).
\item \code{classification_version} Character: describes the version of ClassyFire.
}

If something went wrong, the following placeholders are used for \code{entities}:
all the non-\code{data.frame} elements will be present but assigned
\code{character(0)}. \code{entities$intermediate_nodes} will be present and assigned
an empty list.

The \code{data.frame}s in \code{entities$kingdom}, \code{entities$superclass},
\code{entities$class}, \code{entities$subclass}, and \code{entities$direct_parent} -- and
each of the \code{data.frame}s in the list \code{entities$intermediate_nodes}, if any
-- should have the following variables:
\itemize{
\item \code{name} Character: the name of the classification at this level
\item \code{description} Character: Description of the classification at this level.
\item \code{chemont_id} Character: the ChemOnt node ID of the classification at this level.
\item \code{url} Character: The URL for the ChemOnt node of the classification at this level.
}

If something went wrong, the placeholder for \code{entities$kingdom},
\code{entities$superclass}, \code{entities$class}, \code{entities$subclass}, and
\code{entities$direct_parent} is a \code{data.frame} with the above-listed variables,
but zero rows (i.e., each variable was assigned \code{character(0)}). The
placeholder for \code{entities$intermediate_nodes} is just an empty list.
}

\examples{
# No example

}
\author{
Caroline Ring, Paul Kruse
}
