% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_tip_level}
\alias{get_tip_level}
\title{Tree level}
\usage{
get_tip_level(tree, node)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class)}

\item{node}{Character (a node name) or integer (a node number).}
}
\value{
The level of the node from the root of the tree.
}
\description{
This function returns the tree level of the given node in a rooted tree.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)

get_tip_level(tree = tree, node = 't2') #specify node by name
get_tip_level(tree = tree, node = 2) #specify node by number
\dontshow{\}) # examplesIf}
}
