% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_comparison_dashboard.R
\name{get_toxprints}
\alias{get_toxprints}
\title{A function for retrieving toxprint data for a given chemical}
\usage{
get_toxprints(smiles)
}
\arguments{
\item{smiles}{A string representing a chemical in SMILES format}
}
\value{
A list of chemical descriptors, including ToxPrints descriptors
}
\description{
A function for retrieving toxprint data for a given chemical
}
