% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{handle_missing_node_highlight_clade}
\alias{handle_missing_node_highlight_clade}
\title{Handle missing node highlight clade}
\usage{
handle_missing_node_highlight_clade(
  subtree,
  tree_object,
  list_superclasses,
  tree_visual,
  i,
  color
)
}
\arguments{
\item{subtree}{A \code{phylo} object representing a rooted tree (pruned from a larger tree).}

\item{tree_object}{A \code{phylo} object representing the entire taxonomy from which
parameter \code{subtree} is derived.}

\item{list_superclasses}{A list of superclasses to be highlighted.}

\item{tree_visual}{A \code{ggtree} object that will have clades highlighted.}

\item{i}{The index for which list_superclasses element will be highlighted.}

\item{color}{A color for the clade highlight.}
}
\value{
A ggtree object that will have the specified clade highlighted.
}
\description{
Highlight clades associated with missing nodes
after pruning a tree
}
\details{
This is a helper function to highlight clades associated with missing nodes
after pruning a tree. Missing nodes occur when a node from an original tree
loses all but one child in the pruning process and is thus no longer
considered a node in the subtree.
}
\seealso{
\code{\link{handle_missing_node_show_clade}}
}
\author{
Paul Kruse
}
