% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{handle_missing_node_show_clade}
\alias{handle_missing_node_show_clade}
\title{Handle missing node show clade}
\usage{
handle_missing_node_show_clade(
  subtree,
  tree_object,
  list_superclasses,
  tree_visual,
  i,
  color
)
}
\arguments{
\item{subtree}{A \code{phylo} object representing a rooted tree.}

\item{tree_object}{A \code{phylo} object representing the entire taxonomy from
which the \code{subtree} is derived.}

\item{list_superclasses}{A list of superclasses to be labeled.}

\item{tree_visual}{A \code{ggtree} object that will have clades labeled.}

\item{i}{The index for which list_superclasses element will be labeled.}

\item{color}{A color for the clade label.}
}
\value{
A ggtree object that will have the specified clade labeled.
}
\description{
Display clade labels associated with missing
nodes after pruning a tree
}
\details{
This is a helper function to display clade labels associated with missing
nodes after pruning a tree. Missing nodes occur when a node from an original
tree loses all but one child in the pruning process and is thus no longer
considered a node in the subtree.
}
\seealso{
\code{\link{handle_missing_node_highlight_clade}}
}
\author{
Paul Kruse
}
