% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{HazardComparisonDashboard-class}
\alias{HazardComparisonDashboard-class}
\title{HazardComparisonDashboard class}
\description{
An S4 class to store the results from the Hazard Comparison Dashboard API
}
\section{Slots}{

\describe{
\item{\code{meta}}{A list of the DTXSID, CASRN, name, SMILES, and InChIKey}
}}

