% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{htmap_split_check}
\alias{htmap_split_check}
\title{Heatmap split check}
\usage{
htmap_split_check(
  htmap_split,
  matrix_dim,
  matrix_inds,
  matrix_names,
  tree_object,
  tax_level_labels,
  err_msg = "Invalid heatmap split specification."
)
}
\arguments{
\item{htmap_split}{The split specification (either \code{row_split} or
\code{column_split} as provided to \code{\link[=generate_heatmap]{generate_heatmap()}}).}

\item{matrix_dim}{Integer: The size of the appropriate dimension of the full
matrix to be heatmapped (i.e., argument \code{matrix} as supplied to
\code{\link[=generate_heatmap]{generate_heatmap()}}): \code{nrow(matrix)} for \code{row_split}, \code{ncol(matrix)} for
\code{column_split}.}

\item{matrix_inds}{Integer vector: The row or column indices of the full
matrix to be heatmapped (i.e., argument \code{matrix} as supplied to
\code{\link[=generate_heatmap]{generate_heatmap()}}).}

\item{matrix_names}{Character vector: The row or column names of the full
matrix (i.e., argument \code{matrix} as supplied to \code{\link[=generate_heatmap]{generate_heatmap()}}).}

\item{tree_object}{The tree object from which the matrix was derived (i.e.,
argument \code{tree_object} as supplied to \code{\link[=generate_heatmap]{generate_heatmap()}}). May be a
pruned subset of the original tree (i.e., if \code{\link[=generate_heatmap]{generate_heatmap()}} was
called with a non-\code{NULL} value for argument \code{prune_to}, then this will be
\code{prune_tree(tree = tree_obj, prune_to = prune_to)}).}

\item{tax_level_labels}{Character vector of taxonomy level labels, as
supplied to \code{\link[=generate_heatmap]{generate_heatmap()}}.}

\item{err_msg}{Character: An error message to be thrown if the \code{htmap_split}
specification is not one of the accepted types. Default \code{"Invalid heatmap split specification."}, but \code{\link[=generate_heatmap]{generate_heatmap()}} internally supplies its
own, more-informative error messages.}
}
\value{
If it does not stop with an error, returns the final row/column split
specification. If split specification was \code{"level"} or the name of a
taxonomy level, then the returned object will be a character vector the
same length as \code{matrix_inds}. Otherwise, it will be the same class as
\code{htmap_split}, but with the same length or number of rows as \code{matrix_inds},
unless \code{htmap_split} was a single integer, in which case the returned
object is the same single integer.
}
\description{
A helper function for \code{\link[=generate_heatmap]{generate_heatmap()}} to check the
row/column split specifications.
}
\details{
See \code{\link[=generate_heatmap]{generate_heatmap()}} for possible heatmap split specifications.
}
\examples{
#No examples. See ?generate_heatmap for acceptable split specifications.
}
\author{
Caroline Ring
}
