% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Is whole number}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\description{
Checks whether a number is a whole number (i.e., an integer)
}
\details{
Adapted from the example in the documentation for
\code{\link[base:integer]{base::is.integer()}}.
}
\examples{
is.wholenumber(5) #returns TRUE
is.wholenumber(1.3) #returns FALSE
is.wholenumber(NA_real_) #returns NA
is.wholenumber(numeric(0)) #returns logical(0)
}
