% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{is_inchikey}
\alias{is_inchikey}
\title{Is InChIKey}
\usage{
is_inchikey(x)
}
\arguments{
\item{x}{Character: A vector of one or more strings to check for InChIKey
format.}
}
\value{
A logical vector the same length as \code{x}, containing \code{TRUE} if valid
InChIKey format, and \code{FALSE} otherwise.
}
\description{
Check if a string is a validly-formatted InChIKey
}
\details{
The criteria are the following, as described in Heller et al. (2015)
(see References).
\itemize{
\item must have 27 characters
\item the first 14 characters must be uppercase letters
\item the 15th character must be a hyphen
\item the 16th-23rd characters must be uppercase letters
\item the 24th character must be either "S" or "N"
\item the 25th character must be "A"
\item the 26th character must be a hyphen
\item the 27th character must be an uppercase letter
}
}
\examples{
is_inchikey("PLDWAJLZAAHOGG-UHFFFAOYSA-N") #returns TRUE
is_inchikey("PLDWAJLZAAHOGG-UHFFFAOYSB-N") #returns FALSE
is_inchikey("BAD-INCHIKEY-X") #returns FALSE
is_inchikey(5.234) #returns FALSE
is_inchikey(NA_character_) #returns FALSE
is_inchikey(NULL) #returns FALSE

}
\references{
Heller, S.R., McNaught, A., Pletnev, I. et al. InChI, the IUPAC
International Chemical Identifier. J Cheminform 7, 23 (2015).
https://doi.org/10.1186/s13321-015-0068-4
}
\author{
Caroline Ring
}
