% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Jaccard similarity}
\usage{
jaccard_similarity(tree = NULL, node_1 = NULL, node_2 = NULL)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class)}

\item{node_1}{First node of interest. Character (a node name) or integer (a
node number).}

\item{node_2}{Second node of interest. Character (a node name) or integer (a
node number).}
}
\value{
Numeric: The Jaccard similarity of the ancestry of the two nodes.
}
\description{
This function takes in a tree and two input nodes (either labels or node
numbers) and returns the Jaccard similarity values of the nodes based on the
tree structure. For each node, there is a set of labels along the unique path from the
root to the label. These sets are compared using Jaccard similarity. For more
information on Jaccard similarity, please consult
\href{https://en.wikipedia.org/wiki/Jaccard_index}{Jaccard Index}.
}
\details{
For rapid calculation of similarity values of multiple pairs of nodes,
consider using \code{\link{similarity_matrix}}.
}
\examples{
tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)
jaccard_similarity(tree = tree,
node_1 = 1,
node_2 = 2)


}
\references{
\insertRef{pekar2002taxonomy}{treecompareR}

\insertRef{pesquita2009semantic}{treecompareR}
}
\seealso{
\code{\link{resnik_similarity}}, \code{\link{lin_similarity}},
\code{\link{jiang_conrath_similarity}}, \code{\link{similarity_matrix}}
}
