% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{jiang_conrath_similarity}
\alias{jiang_conrath_similarity}
\title{Jiang Conrath Similarity}
\usage{
jiang_conrath_similarity(tree = NULL, node_1 = NULL, node_2 = NULL)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class)}

\item{node_1}{First node of interest. Character (a node name) or integer (a
node number).}

\item{node_2}{Second node of interest. Character (a node name) or integer (a
node number).}
}
\value{
The Jiang Conrath similarity value between the two input nodes.
}
\description{
This function takes in a tree and two input nodes (either labels or node
numbers) and returns the Jiang and Conrath similarity values of the nodes
based on the tree structure. The function uses the formulation as described
in
\href{https://www.researchgate.net/publication/220837848_An_Intrinsic_Information_Content_Metric_for_Semantic_Similarity_in_WordNet/stats}{An
Intrinsic Information Content Metric for Semantic Similarity in WordNet}.
Furthermore, this function is a wrapper for a RCPP function. For rapid
calculation of similarity values of several pairs of nodes, consider using
\code{\link{similarity_matrix}}.
}
\references{
\insertRef{seco2004intrinsic}{treecompareR}

\insertRef{jiang1997semantic}{treecompareR}
}
\seealso{
\code{\link{jaccard_similarity}}, \code{\link{resnik_similarity}},
\code{\link{lin_similarity}}, \code{\link{similarity_matrix}}
}
