% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{parse_item}
\alias{parse_item}
\title{Parse list item}
\usage{
parse_item(entities, item, query_type, tax_level_labels = chemont_tax_levels)
}
\arguments{
\item{entities}{A \code{data.frame} of classified entities from one page of
ClassyFire results, e.g, the \code{entities} element of one element of the list
output by \code{\link[=query_structure]{query_structure()}}.}

\item{item}{Character: The name of an item in \code{entities}. One of \code{'alternative_parents'},
\code{'molecular_framework'}, \code{'substituents'}, \code{'description'},
\code{'external_descriptors'}, \code{'ancestors'}, \code{'predicted_chebi_terms'},
\code{'predicted_lipidmaps_terms'}. See Details.
These are all the named items currently in \code{entities} as returned by
ClassyFire, so providing any other string here will result in an empty
\code{data.frame} being returned.}
}
\value{
A \code{data.frame} containing the information from the specified item.
Contains variables \code{identifier}, \code{smiles}, \code{inchikey},
\code{classification_version}, and other variables depending on what item was
specified. If \code{item = 'alternative_parents'}, additional variables are
\code{name}, \code{description}, \code{chemont_id}, \code{url}. For any of the other possible
values of \code{item} listed above, there will be
one additional variable named for \code{item} (e.g., if \code{item = 'molecular_framework'}, the additional variable will be named
\code{molecular_framework}). If
\code{item} is some string other than those listed above, i.e., not matching any
of the named items in \code{entities}, then an empty \code{data.frame} will be
returned (zero rows and no variables).
}
\description{
Parse an item from a list of classified entities.
}
\details{
This is a helper function used by \code{\link[=classify_structures]{classify_structures()}} and
\code{\link[=classify_inchikeys]{classify_inchikeys()}}. It generally should not be called directly by the
user.

The main classification is parsed by \code{\link[=parse_classification]{parse_classification()}}, but
ClassyFire returns several additional pieces of classification output.
These include:
\itemize{
\item Levels of classification: \code{kingdom}, \code{superclass}, \code{class}, \code{subclass}, \code{intermediate_nodes}, \code{direct_parent}
\item Alternative parents (\code{alternative_parents})
\item Molecular framework (\code{molecular_framework})
\item Substituents (\code{substituents})
\item Description (\code{description})
\item External descriptors (\code{external_descriptors})
\item Ancestors (\code{ancestors})
\item Predicted ChEbI terms (\code{predicted_chebi_terms})
\item Predicted LipidMaps terms (\code{predicted_lipidmaps_terms})

This function parses any of these additional items into a \code{data.frame}.

The resulting \code{data.frame} will have variables \code{identifier}, \code{smiles},
\code{inchikey}, \code{classification_version}, and the following additional
variables depending on what item was requested:
\item \code{kingdom}, \code{superclass}, \code{class}, \code{subclass}, \code{intermediate_nodes}, \code{direct_parent}, and \code{alternative_parents}: \code{name}, \code{description}, \code{chemont_id}, \code{url}
\item \code{external_descriptors}: If any entities had these, \code{source} (the external source, e.g. "CHEBI", "KEGG", or "LIPID MAPS"); \code{source_id} (the ID of the descriptor in the external source, e.g. "CHEBI:15756" from ChEBI, or "LMFA01010001" from LipidMaps); and \code{annotations}, the annotation of the external descriptor (e.g. "long-chain fatty acid" or "Straight chain fatty acids")
\item everything else: a single additional variable named after the requested item (i.e., \code{molecular_framework}, \code{substituents}, etc.)
}
}
\examples{
#get a set of results to parse
my_results <- query_structure(input = c("COC1=CC(Br)=CC=C1", "SCCSCCS"))
#get alternative parents
parse_item(entities = my_results[[1]]$entities, item = "alternative_parents")
#get molecular framework
parse_item(entities = my_results[[1]]$entities, item = "molecular_framework")
}
\author{
Caroline Ring
}
