% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{prune_and_display_subtree}
\alias{prune_and_display_subtree}
\title{Prune and display subtree}
\usage{
prune_and_display_subtree(
  base_tree = chemont_tree,
  base_name = NULL,
  prune_to = NULL,
  prune_name = NULL,
  adjust_branch_length = FALSE,
  tax_level_labels = chemont_tax_levels,
  ...
)
}
\arguments{
\item{adjust_branch_length}{An alternate parameter determining whether to
resize branches of subtree.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}

\item{data}{A data.table with classification data for chemicals.}

\item{tree}{An alternate parameter giving a taxonomy if not using ChemOnt.}

\item{show_tips}{An alternate parameter determining whether tip labels are
displayed.}

\item{no_plot}{An alternate parameter for returning only the pruned tree
without the tree visual.}

\item{xlimit}{An alternate parameter determining the scale of the tree
visualization.}
}
\value{
A pruned tree or list consisting of a pruned tree and ggtree diagram
of the pruned tree.
}
\description{
This function takes a data.table of chemicals and their classification data,
and returns the data-induced subtree with an optional tree diagram of the
subtree.
}
