% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{prune_c}
\alias{prune_c}
\title{Prune concatenate}
\usage{
prune_c(prune_to)
}
\value{
If \code{prune_to} was a list, returns an object of the same class as the
list elements, consisting of the concatenation of the list elements (with
\code{\link[=unique]{unique()}} applied). If \code{prune_to} was not a list, then returns \code{prune_to}
unchanged.
}
\description{
Concatenate a list of pruning specifications
}
\details{
If pruning specification is supplied as a list item to
\code{\link[=prune_tree]{prune_tree()}}, concatenate the list and return. This is a helper function
for \code{\link[=prune_tree]{prune_tree()}} and generally should not be called directly by the user.
}
\examples{
prune_c(prune_to = list(biosolids_class[1:10, ], usgs_class[1:10, ]))



}
\author{
Caroline Ring
}
