% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{query_classyfire_inchikey}
\alias{query_classyfire_inchikey}
\title{Query ClassyFire InChIKey}
\usage{
query_classyfire_inchikey(inchikey, retry_get_times = 3, wait_min = 5)
}
\arguments{
\item{inchikey}{A character string encoding the InChIKey of the chemical
under query.}

\item{retry_get_times}{The number of times to retry the query, a positive
integer with default value 3.}

\item{wait_min}{The number of seconds to pause in between attempts, used in
the \code{\link{httr}{RETRY}} function.}
}
\value{
A JSON file with the classification data.
}
\description{
This is a helper function that is used to communicate with the ClassyFire API
in order to retrieve classifications.
}
