% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{query_inchikey}
\alias{query_inchikey}
\title{Query ClassyFire by InChIKey}
\usage{
query_inchikey(
  inchikey,
  retry_get_times = 3,
  wait_sec = 5,
  terminate_on = c(400:407, 409:418, 421:428, 431, 451, 500:511),
  base_url = "http://classyfire.wishartlab.com/entities"
)
}
\arguments{
\item{inchikey}{Character: one InChIKey to be queried.}

\item{retry_get_times}{Integer: The number of times to retry the query before
giving up if it returns an HTTP error. Default 3.}

\item{wait_sec}{Numeric: The number of seconds to pause between retry
attempts. Default 5, because ClassyFire rate-limits to 12 requests per
minute. If set to less than 5, it will be reset to 5, with a message.}

\item{base_url}{The base URL for ClassyFire API queries. Default
\code{"http://classyfire.wishartlab.com/entities"}. If you change this, the query
is highly unlikely to work!}
}
\value{
A list consisting of the parsed JSON for each page of results. A completed
query will have named elements \code{id}, \code{query_url}, \code{query_status}, \code{label},
\code{classification_status}, \code{number_of_elements}, \code{number_of_pages},
\code{invalid_entities}, and \code{entities}. A failed query will have named elements
\code{id}, \code{query_url}, \code{query_status}, \code{label}, \code{classification_status}, and
\code{number_of_pages}. \code{id} gives the numerical query ID (assigned by
ClassyFire). \code{query_url} gives the queried URL. \code{label} gives the
user-supplied query label. \code{query_status} gives the HTTP status of the
request (200 means successful). \code{classification_status} reports the
ClassyFire status: "Done" means classification was successfully completed;
"In Queue" means the query timed out while it was still queued; "In
Progress" means the query timed out while it was still processing; "Failed"
means the query failed before ClassyFire could report a status (e.g. HTTP
status code other than 200, or ClassyFire returned some content that did not
include a classification status). \code{number_of_elements} gives the number of
classified entities. \code{number_of_pages} gives the total number of pages for
this query. \code{invalid_entities}, if present, is a \code{data.frame} listing queried
entity identifiers that ClassyFire found invalid. \code{entities}, if present, is
a nested \code{data.frame} giving classifications.
}
\description{
Query ClassyFire API by InChIKey.
}
\details{
This is a helper function that used by \code{\link[=classify_inchikeys]{classify_inchikeys()}}. It
should generally not be called directly by the user.
}
\examples{
query_inchikey(inchikey = "PLDWAJLZAAHOGG-UHFFFAOYSA-N")

}
