% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{query_structure}
\alias{query_structure}
\title{Query ClassyFire by structure}
\usage{
query_structure(
  input = NULL,
  url = NULL,
  label = "query",
  type = "STRUCTURE",
  retry_get_times = 10,
  wait_sec = 5,
  retry_query_times = 10,
  processing_wait_per_input = 0.1,
  terminate_on = c(400:407, 409:418, 421:428, 431, 451, 500:511),
  base_url = "http://classyfire.wishartlab.com/queries"
)
}
\arguments{
\item{input}{A character vector of structural identifiers: SMILES strings or
InChi strings. May optionally be named. Default NULL to use \code{url} instead.
Either \code{input} or \code{url} must be provided.}

\item{url}{Optional: a string giving an existing ClassyFire query URL to
retry. Default NULL to use \code{input} instead. Either \code{input} or \code{url} must be
provided.}

\item{label}{An optional text label for the ClassyFire query. Default
\code{"query"}.}

\item{type}{String giving ClassyFire query type. Default \code{"structure"}.}

\item{retry_get_times}{Max number of times to retry GET command to ClassyFire
API to get status of query (with wait time in between tries). Default 10.}

\item{wait_sec}{Minimum number of seconds to wait before retrying any GET
command or query. Default 5 seconds (because ClassyFire rate-limits
requests to 12 per minute). If less than 5 seconds, will be reset to 5
seconds, with a warning. Wait times for multiple retries
use exponential backoff with full jitter (wait time is \code{runif(1,
  wait_sec, wait_sec * 2^(attempt_number)}).}

\item{retry_query_times}{Max number of attempts to retry an "In Queue" or
"Processing" query (with wait time in between tries). Default 10.}

\item{processing_wait_per_input}{Minimum number of seconds to wait \emph{per input
string} before retrying to retrieve results for a query whose status is
"Processing." Default 0.1. Uses exponential backoff: effective wait time
per input = \code{processing_wait_per_input * 2^(attempt number)}. Total
wait time is either the effective wait time per input times the number of
input structures, or \code{wait_sec} seconds, whichever is greater. (This
means \code{processing_wait_per_input} effectively can never be less than
\code{wait_sec/length(input)}, even if you set it to a smaller number.)}

\item{terminate_on}{Integer vector: List of HTTP status codes which will
immediately terminate with no more retries. Default: \code{ c(400:407, 409:418, 421:428, 431, 451, 500:511)}}

\item{base_url}{The base URL for ClassyFire API queries. Default
\code{"http://classyfire.wishartlab.com/entities"}. If you change this, the
query is highly unlikely to work!}
}
\value{
A list of lists. The outer list has one element for each page of the
JSON output from the ClassyFire query (there is one page for every ten
entities). If the ClassyFire query failed or timed out, the list will have
one element. Each list element is itself a list, consisting of the parsed
JSON result for each page. A completed query will have named elements \code{id},
\code{query_url}, \code{query_status}, \code{label}, \code{classification_status},
\code{number_of_elements}, \code{number_of_pages}, \code{invalid_entities}, and
\code{entities}. A failed query will have named elements \code{id}, \code{query_url},
\code{query_status}, \code{label}, \code{classification_status}, and \code{number_of_pages}.
\code{id} gives the numerical query ID (assigned by ClassyFire). \code{query_url}
gives the queried URL. \code{label} gives the user-supplied query label.
\code{query_status} gives the HTTP status of the request (200 means successful).
\code{classification_status} reports the ClassyFire status: "Done" means
classification was successfully completed; "In Queue" means the query timed
out while it was still queued; "In Progress" means the query timed out
while it was still processing; "Failed" means the query failed before
ClassyFire could report a status (e.g. HTTP status code other than 200, or
ClassyFire returned some content that did not include a classification
status). \code{number_of_elements} gives the number of classified entities.
\code{number_of_pages} gives the total number of pages for this query.
\code{invalid_entities}, if present, is a \code{data.frame} listing queried entity
identifiers that ClassyFire found invalid. \code{entities}, if present, is a
nested \code{data.frame} giving classifications.

@examples
query_structure(input = c("COC1=CC(Br)=CC=C1", "SCCSCCS"))
}
\description{
This function takes a vector of structural identifiers (SMILES strings or
InChi strings) and queries the ClassyFire API to get classifications for each
one.
}
\details{
See ClassyFire API details at http://classyfire.wishartlab.com/access.
}
