% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{resnik_similarity}
\alias{resnik_similarity}
\title{Resnik Similarity}
\usage{
resnik_similarity(tree = NULL, node_1 = NULL, node_2 = NULL)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class)}

\item{node_1}{First node of interest. Character (a node name) or integer (a
node number).}

\item{node_2}{Second node of interest. Character (a node name) or integer (a
node number).}
}
\value{
The Resnik similarity value between the two input nodes.
}
\description{
This function takes in a tree and two input nodes (either labels or node
numbers) and returns the Resnik similarity values of the nodes based on the
tree structure. The function uses the formulation as described in
\href{https://www.researchgate.net/publication/220837848_An_Intrinsic_Information_Content_Metric_for_Semantic_Similarity_in_WordNet/stats}{An
Intrinsic Information Content Metric for Semantic Similarity in WordNet}. For
rapid calculation of similarity values of several pairs of nodes, consider
using \code{\link{similarity_matrix}}.
}
\examples{
tree <- generate_topology(n = 8, rooted = TRUE, seed = 19)
resnik_similarity(tree = tree,
node_1 = 1,
node_2 = 2)


}
\references{
\insertRef{lin1998information}{treecompareR}

\insertRef{resnik1995using}{treecompareR}
}
\seealso{
\code{\link{jaccard_similarity}}, \code{\link{lin_similarity}},
\code{\link{jiang_conrath_similarity}}, \code{\link{similarity_matrix}}
}
