% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_tree.R
\name{select.tip.or.node}
\alias{select.tip.or.node}
\title{Get label for a tip or internal node of a phylo tree}
\usage{
select.tip.or.node(element, tree)
}
\arguments{
\item{element}{Vector of node numbers in phylo tree}

\item{tree}{phylo tree object}
}
\value{
Character vector of tip or internal node labels
}
\description{
Get label for a tip or internal node of a phylo tree
}
