% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_tree.R
\name{SetPlotHeight}
\alias{SetPlotHeight}
\title{Helper function for fast tree construction}
\usage{
SetPlotHeight(node, rootHeight = 100)
}
\arguments{
\item{node}{A data.tree object.}

\item{rootHeight}{A positive number placeholder for root height, default 100.}
}
\value{
A data.tree object with plotHeight attribute.
}
\description{
Helper function for fast tree construction
}
