% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{similarity_matrix}
\alias{similarity_matrix}
\title{Similarity matrix}
\usage{
similarity_matrix(
  tree = NULL,
  nodes1 = NULL,
  nodes2 = NULL,
  metric = "jaccard",
  upper_tri = TRUE
)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class)}

\item{nodes1}{A vector of the first set of nodes of interest. Character (node
names) or integer (node numbers). If \code{NULL} (default), will be taken as all
nodes in \code{tree}.}

\item{nodes2}{A vector of the second set of nodes of interest. Character
(node names) or integer (node numbers). If \code{NULL} (default), will be taken as all
nodes in \code{tree}.}

\item{metric}{A string naming the similarity metric to use. Options are
"jaccard", "resnik", "lin", and "jiang_conrath". Only the first two letters
need be entered (e.g., "ja" for "jaccard", "re" for "resnik", "li" for
"lin", and "ji" for "jiang_conrath"). Case-insensitive.}

\item{upper_tri}{Logical: If TRUE (default), only the upper triangular part
of the similarity matrix is constructed and returned. If FALSE, the full
matrix is constructed and returned.}
}
\value{
A similarity matrix. The dimension names for the rows and columns
correspond to the associated node labels of the tree.
}
\description{
This function takes in a list of node labels for rows and for columns (or
node indices), a tree, and a similarity measure, and returns a matrix with
the calculated similarity values for each pair. Note that the default is to
compute the similarity values only for pairs along and above the diagonal, so
when the set of nodes for rows and columns differ, one must turn off this
feature.
}
\details{
To calculate the similarity matrix for all pairs of nodes in a tree, provide
\code{tree} and leave \code{labels1}, \code{labels2}, \code{nodes1}, and \code{nodes2} all \code{NULL}.
}
\seealso{
\code{\link{jaccard_similarity}}, \code{\link{resnik_similarity}},
\code{\link{lin_similarity}}, \code{\link{jiang_conrath_similarity}}
}
