% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{simulate_trees}
\alias{simulate_trees}
\title{Tree simulator}
\usage{
simulate_trees(n_trees = 1, n, ...)
}
\arguments{
\item{n_trees}{Integer: The number of trees to generate. Default 1}

\item{n}{The number of tips each tree will have.}

\item{...}{Parameters passed to the generate_topology() function.}
}
\value{
A list of 'phylo' objects, each representing a simulated tree. (See
\code{\link[ape:read.tree]{ape::read.tree()}} for description of the \code{phylo} class.)
}
\description{
Randomly generate trees.
}
\details{
This function generates several trees subject to input constraints.
}
\examples{
set.seed(42)
simulate_trees(n_trees = 2, n = 5) #a list of two trees, plot each using ggtree()

}
\author{
Paul Kruse
}
