% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{terminal_function}
\alias{terminal_function}
\title{Helper function to return the terminal label for \code{add_terminal_label()}}
\usage{
terminal_function(t, tip = FALSE, tax_level_labels, tree = NULL)
}
\arguments{
\item{t}{A data.table with classification data.}

\item{tip}{Alternate parameter for determining whether to only return tip
labels.}

\item{tax_level_labels}{Parameter giving classification levels.}

\item{tree}{Alternate parameter for giving a taxonomy tree different from
ChemOnt.}
}
\value{
A string giving the terminal label.
}
\description{
Helper function to return the terminal label for \code{add_terminal_label()}
}
