% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usgs}
\alias{usgs}
\title{usgs chemical list from the US EPA CompTox Chemicals Dashboard}
\format{
A csv representing 707 rows and 18 columns.
\describe{
\item{DTXSID}{A chemical identifier}
\item{PREFERRED.NAME}{A character string of the name}
\item{CASRN}{A character string of the CASRN identifier}
\item{INCHIKEY}{A character of the InChIKey identifier}
\item{IUPAC.NAME}{A character string of the IUPAC identifier}
\item{SMILES}{A character string of the SMILES identifier}
\item{INCHI.STRING}{A character giving the INCHI string identifier.}
\item{MOLECULAR.FORMULA}{A character string of the molecular formula}
\item{AVERAGE.MASS}{A numeric giving the average atomic mass}
\item{MONOISOTOPIC.MASS}{A numeric giving the monoisotopic mass}
\item{Num.of.SOURCES}{An integer giving the number of sources}
\item{Num.of.PUBMED.ARTICLES}{An integer giving the number of PUBMED
Articles}
\item{PUBCHEM.DATA.SOURCES}{An integer giving the number of PUBCHEM data
sources}
\item{CPDAT.COUNT}{An integer giving the number of occurrences within CPDAT}
\item{QC.Level}{An integer giving the quality control level}
\item{Num.ToxCast.Active}{An integer: Number of ToxCast assays that showed bioactivity}
\item{Pct.ToxCast.Active}{A numeric: Percent of ToxCast assays that showed bioactivity}
\item{Total.Assays}{An integer: Total number of ToxCast assays in which the chemical was tested}}
}
\usage{
usgs
}
\description{
A data set containing a list of 707 chemicals with related information.
}
\details{
\url{https://comptox.epa.gov/dashboard/chemical-lists/USGSWATER}
}
\keyword{datasets}
