% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usgs_class}
\alias{usgs_class}
\title{usgs chemical list from the US EPA CompTox Chemicals Dashboard}
\format{
A csv representing 688 rows and 31 columns.
\describe{
\item{INCHIKEY}{A character of the InChIKey identifier}
\item{DTXSID}{A chemical identifier}
\item{PREFERRED.NAME}{A character string of the name}
\item{CASRN}{A character string of the CASRN identifier}
\item{IUPAC.NAME}{A character string of the IUPAC identifier}
\item{SMILES}{A character string of the SMILES identifier}
\item{INCHI.STRING}{A character giving the INCHI string identifier.}
\item{MOLECULAR.FORMULA}{A character string of the molecular formula}
\item{AVERAGE.MASS}{A numeric giving the average atomic mass}
\item{MONOISOTOPIC.MASS}{A numeric giving the monoisotopic mass}
\item{Num.of.SOURCES}{An integer giving the number of sources}
\item{Num.of.PUBMED.ARTICLES}{An integer giving the number of PUBMED
Articles}
\item{PUBCHEM.DATA.SOURCES}{An integer giving the number of PUBCHEM data
sources}
\item{CPDAT.COUNT}{An integer giving the number of occurrences within CPDAT}
\item{QC.Level}{An integer giving the quality control level}
\item{Num.ToxCast.Active}{An integer: Number of ToxCast assays that showed bioactivity}
\item{Pct.ToxCast.Active}{A numeric: Percent of ToxCast assays that showed bioactivity}
\item{Total.Assays}{An integer: Total number of ToxCast assays in which the chemical was tested}
\item{kingdom}{A character string giving the ClassyFire kingdom label}
\item{superclass}{A character string giving the ClassyFire superclass label}
\item{class}{A character string giving the ClassyFire class label}
\item{subclass}{A character string giving the ClassyFire subclass label}
\item{level5}{A character string giving the ClassyFire level5 label}
\item{level6}{A character string giving the ClassyFire level6 label}
\item{level7}{A character string giving the ClassyFire level7 label}
\item{level8}{A character string giving the ClassyFire level8 label}
\item{level9}{A character string giving the ClassyFire level9 label}
\item{level10}{A character string giving the ClassyFire level10 label}
\item{level11}{A character string giving the ClassyFire level11 label}
\item{terminal_label}{A character string giving the most specific ClassyFire
label for classification}
\item{terminal_level}{An integer giving the level of the \code{terminal_label}}
}
}
\usage{
usgs_class
}
\description{
A data set containing a list of 707 chemicals with related information.
}
\details{
\url{https://comptox.epa.gov/dashboard/chemical-lists/USGSWATER}
}
\keyword{datasets}
