% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USGSWATER_class}
\alias{USGSWATER_class}
\title{USGSWATER chemical list with ClassyFire classifications}
\format{
A csv representing 707 rows and 18 columns \describe{ \item{DTXSID}{A
chemical identifier} \item{PREFERRED_NAME}{A character string of the name}
\item{CASRN}{A character string of the CASRN identifier} \item{INCHIKEY}{A
character of the InChIKey identifier} \item{SMILES}{A character string of
the SMILES identifier} }
}
\usage{
USGSWATER_class
}
\description{
A data set containing a list of 707 chemicals with related information
}
\details{
Original data source:
\url{https://comptox.epa.gov/dashboard/chemical-lists/USGSWATER}. ClassyFire
classifications obtained via ClassyFire API.
}
\seealso{
\code{\link{USGSWATER}}
}
\keyword{datasets}
