% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{add_terminal_label}
\alias{add_terminal_label}
\title{This function returns the terminal classification label for classified chemicals}
\usage{
add_terminal_label(data, tip = FALSE, labels = NULL, tree = NULL)
}
\arguments{
\item{data}{A data.table of classified chemicals.}

\item{tip}{An alternate parameter controlling whether to include only tip labels.}

\item{labels}{An alternate parameter giving a list of the classification levels when not using ClassyFire classified data.}

\item{tree}{An alternate parameter giving a different tree structure if not using ChemOnt taxonomy.}
}
\value{
A new data.table, augmenting the input data.table with a column consisting of the terminal labels.
}
\description{
This function returns the terminal classification label for classified chemicals
}
