% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{add_terminal_label}
\alias{add_terminal_label}
\title{Add terminal label}
\usage{
add_terminal_label(data, tax_level_labels = chemont_tax_levels)
}
\arguments{
\item{data}{A data.frame of classified chemicals.}

\item{tax_level_labels}{A vector of taxonomy levels. Default is
\code{\link{chemont_tax_levels}} to use the levels of the ChemOnt taxonomy.}

\item{tree}{An alternate parameter giving a different tree structure if not
using ChemOnt taxonomy.}
}
\value{
A new data.frame, augmenting the input data.frame with a column
consisting of the terminal labels.
}
\description{
This function takes in a data.frame of classified entities and determines the
terminal classification label for each entity.
}
