% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{adjust_branch_lengths}
\alias{adjust_branch_lengths}
\title{Adjust branch lengths}
\usage{
adjust_branch_lengths(tree)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class).}
}
\value{
Numeric: A vector of adjusted branch lengths for the input tree.
}
\description{
Adjust branch lengths for pruned trees.
}
\details{
This is a helper function used by \code{\link[=prune_tree]{prune_tree()}}. Usually it will not
be called by the user directly.

This function adjusts the length of branches after pruning. With the
adjustment in this function, internal nodes that have become tip nodes after
pruning (\emph{i.e.}, because their descendants have been pruned away) will be
plotted at the same branch length as other tip nodes.
}
\examples{
my_tree <- prune_tree(tree = chemont_tree,
 prune_to = biosolids_class[1:10,])
adjust_branch_lengths(tree = tree)

}
\author{
Paul Kruse, Caroline Ring
}
