% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{adjust_branch_lengths}
\alias{adjust_branch_lengths}
\title{Adjust branch lengths}
\usage{
adjust_branch_lengths(tree)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}
}
\value{
A tree with adjust branch lengths.
}
\description{
This is a helper function that is used to reset branch lengths for pruned
trees.
}
