% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{as_classified.phylo}
\alias{as_classified.phylo}
\title{Convert a phylo tree into a wide-format "classified" data.frame}
\usage{
as_classified.phylo(
  tree,
  tax_level_labels = c("kingdom", "superclass", "class", "subclass", "level5", "level6",
    "level7", "level8", "level9", "level10", "level11")
)
}
\arguments{
\item{tree}{An object of class \code{phylo}.}

\item{tax_level_labels}{Vector of the possible taxonomy levels that can appear
as column names in \code{as_classified.phylo} if it is a \code{data.frame} of
classified data.}
}
\description{
Convert a phylo tree into a wide-format "classified" data.frame
}
