% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{attach_information_content}
\alias{attach_information_content}
\title{Attach information content}
\usage{
attach_information_content(tree, log_descendants = TRUE)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class)}

\item{log_descendants}{Alternate parameter determining type of information
content to use.}
}
\value{
phylo object with information content data.frame attached
}
\description{
This function attaches information content if missing for input tree. This
uses the formulation as described in
\href{https://www.researchgate.net/publication/220837848_An_Intrinsic_Information_Content_Metric_for_Semantic_Similarity_in_WordNet/stats}{An
Intrinsic Information Content Metric for Semantic Similarity in WordNet}.
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)

tree <- attach_information_content(tree = tree)
\dontshow{\}) # examplesIf}
}
\references{
\insertRef{seco2004intrinsic}{treecompareR}
}
