% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{attach_information_content}
\alias{attach_information_content}
\title{Attaches information content if missing for input tree}
\usage{
attach_information_content(tree, log_descendants = TRUE)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{log_descendants}{Alternate parameter determining type of information content to use.}
}
\value{
phylo object with information content data.frame attached
}
\description{
Attaches information content if missing for input tree
}
