% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{bind_entities}
\alias{bind_entities}
\title{Bind entities to a tree}
\usage{
bind_entities(
  tree,
  data,
  entity_id_col = NULL,
  tax_level_labels = chemont_tax_levels
)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class). Tips will be bound to
this tree.}

\item{data}{Either one \code{data.frame}, or a list of \code{data.frames}, containing
classified entities. The data.frames must include the column names
specifiedin \code{tax_level_labels} and \code{entity_id_col}.}

\item{entity_id_col}{Character vector: One or more variable name(s) in
\code{data} that, together, uniquely identify the entities. Default \code{NULL},
which treats each row in \code{data} as a unique entity named \code{entity1},
\code{entity2}, ... for as many rows as there are in the data. If a vector of
variable names is provided, entities will be named by concatenating rows of
the specified variables.}

\item{tax_level_labels}{Taxonomy levels used for classification in
\code{data}. Default is \link{chemont_tax_levels}.}
}
\value{
A \code{phylo}-class object.
}
\description{
Bind individual classified entities as new tips to a tree
}
\details{
Given a \code{data.frame} of classified entities with identifiers, this
function gets the terminal classifications for each entity, and then binds
the entity ID to each terminal classification as a new tip node of the
tree.

This function may be useful when you plotting classified entities: for
example, if you are using the ChemOnt taxonomy and plotting classified
chemicals, and you want to show all of the individual chemicals with each
classification in the tree, you could use \code{\link[=bind_entities]{bind_entities()}} first and then
plot the resulting tree using, for example, \code{\link[=display_subtree]{display_subtree()}}.
}
\examples{

#explicitly specifying entity ID
bind_entities(tree = prune_tree(tree = chemont_tree,
prune_to = biosolids_class[1:20, ]),
data = biosolids_class[1:20, ],
entity_id_col = "DTXSID")

#not specifying entity ID
bind_entities(tree = prune_tree(tree = chemont_tree,
prune_to = biosolids_class[1:20, ]),
data = biosolids_class[1:20, ])

}
\author{
Caroline Ring, Paul Kruse
}
