% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{bind_entities}
\alias{bind_entities}
\title{Bind individual entities as new tips to a tree}
\usage{
bind_entities(
  tree,
  data,
  entity_id_col,
  tax_level_labels = c("kingdom", "superclass", "class", "subclass", "level5", "level6",
    "level7", "level8", "level9", "level10", "level11")
)
}
\arguments{
\item{tree}{The base tree as a \code{phylo}-class object. Tips will be bound
to this tree.}

\item{data}{Either one data.frame, or a list of data.frames, containing
classified entities. Each row of the data.frame is one entity. The
data.frames must include the column names specifeid in
\code{tax_level_labels} and \code{entity_id_col}.}

\item{entity_id_col}{The column name in \code{data} containing identifying
labels for the entities.}

\item{tax_level_labels}{Taxonomy levels used for classification in
\code{data}. Default is the Chemont taxonomy levels: \code{c('kingdom',
  'superclass', 'class', 'subclass','level5', 'level6', 'level7',
  'level8','level9', 'level10', 'level11')}.}
}
\value{
A \code{phylo}-class object.
}
\description{
Bind individual entities as new tips to a tree
}
