% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{calc_number_overlap}
\alias{calc_number_overlap}
\title{calculate overlap between two classified datasets at the individual entity level}
\usage{
calc_number_overlap(
  data_1,
  data_2,
  entity_id_col,
  at_level = "terminal",
  tax_level_labels = chemont_tax_levels
)
}
\arguments{
\item{data_1}{A data.frame of classified entities to be used as the base set}

\item{compare_dat}{A data.frame of classified entitites to be used as hte comp}
}
\description{
calculate overlap between two classified datasets at the individual entity level
}
