% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{calc_similarity}
\alias{calc_similarity}
\title{Calculate similarity measures}
\usage{
calc_similarity(tree, labels_A, labels_B, metric = "jaccard", method = "cpp")
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{labels_A}{The first (set of) label(s).}

\item{labels_B}{The second (set of) label(s).}
}
\value{
Matrix of similarity values for pairs of labels from \code{label_A} and \code{label_B}.
}
\description{
Calculate similarity measures
}
\examples{

tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)

calc_similarity(tree = tree, labels_A = 't2', labels_B = 't4')
calc_similarity(tree = tree, labels_A = 'n3', labels_B = 't8')

}
\references{
\insertRef{pekar2002taxonomy}{treecompareR}

\insertRef{pesquita2009semantic}{treecompareR}
}
