% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{calc_similarity_data}
\alias{calc_similarity_data}
\title{Calculate similarity measures for two datasets}
\usage{
calc_similarity_data(
  data_1,
  data_2,
  terminal_label = "terminal_label",
  tree = chemont_tree,
  tax_level_labels = chemont_tax_levels,
  similarity = "jaccard"
)
}
\arguments{
\item{data_1}{A data.frame of classified entities}

\item{data_2}{Another data.frame of classified entities}

\item{terminal_label}{The variable name in the two data.frames that denotes
the terminal label of the classification. Default "terminal_label".}

\item{tree}{The taxonomy tree to use. Default \code{\link{chemont_tree}}.}

\item{tax_level_labels}{The set of taxonomy levels to use. Default \code{\link{chemont_tax_levels}}.}

\item{similarity}{The similarity metric to calculate. Default "jaccard".}
}
\description{
Calculate similarity measures for two datasets
}
