% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{check_similarity_inputs}
\alias{check_similarity_inputs}
\title{Check similarity inputs}
\usage{
check_similarity_inputs(
  tree = NULL,
  label_1 = NULL,
  label_2 = NULL,
  node_1 = NULL,
  node_2 = NULL
)
}
\arguments{
\item{tree}{An object of class \code{phylo}.}

\item{label_1}{A label for a node within the \code{tree} object.}

\item{label_2}{A label for a node within the \code{tree} object.}

\item{node_1}{An alternate parameter for a node within the \code{tree} object.}

\item{node_2}{An alternate parameter for a node within the \code{tree} object.}
}
\value{
A pair of node numbers corresponding to the input parameters
specifying the first and second nodes.
}
\description{
This is a helper function for checking user input values within the
\code{\link{jaccard_similarity}}, \code{\link{resnik_similarity}},
\code{\link{lin_similarity}}, \code{\link{jiang_conrath_similarity}}, and
\code{link{similarity_matrix}} functions.
}
