% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{choose_partition}
\alias{choose_partition}
\title{Choose a partition}
\usage{
choose_partition(n, max_deg = NULL, min_deg = NULL)
}
\arguments{
\item{n}{The integer for which a permuted partition is generated.}

\item{max_deg}{An alternate parameter, giving a constraint on the permissible
partitions.}

\item{min_deg}{An alternate parameter, giving a constraint on the permissible
partitions.}
}
\value{
A permuted partition of the integer \code{n}.
}
\description{
This function selects a permuted partition given a positive integer input.
This is a helper function for generating a tree topology closer to random
than \code{generate_topology} would otherwise.
}
