% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{circ_tree_boxplot}
\alias{circ_tree_boxplot}
\title{Circular tree with boxplots}
\usage{
circ_tree_boxplot(data, col, tax_level_labels = NULL, tree = NULL)
}
\arguments{
\item{data}{A data.table consisting of classification data and additional
numeric data.}

\item{col}{A specified numeric column for use in displaying boxplots.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}

\item{tree}{An alternate parameter giving a taxonomy if not using ChemOnt.}
}
\value{
A ggtree object consisting of subtree induced by data and boxplots
corresponding to specified numeric column of data.
}
\description{
This function takes in a data.table of chemicals with classification data and
additional numeric data, and displays selected numeric data grouped by tip
label on the data-induced subtree of the classification taxonomy.
}
