% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{circ_tree_boxplot}
\alias{circ_tree_boxplot}
\title{Circular tree with boxplots}
\usage{
circ_tree_boxplot(
  data,
  col,
  tax_level_labels = chemont_tax_levels,
  title = NULL,
  tree = chemont_tree,
  layers = NULL,
  adjust_branch_length = FALSE,
  tippoint_boxplot = FALSE
)
}
\arguments{
\item{data}{A data.table consisting of classification data and additional
numeric data.}

\item{col}{A specified numeric column for use in displaying boxplots.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}

\item{title}{An alternate parameter for the title of the plot.}

\item{tree}{An alternate parameter giving a taxonomy if not using ChemOnt.}

\item{layers}{An alternate parameter giving which taxonomic layers to display
outside of the boxplot layer. This can be either a string with a single
column name, a vector of column names, or a list of column names. If the
input is a vector or a list, it is fine for it to be length 1.}

\item{adjust_branch_length}{Whether to adjust branch length so that all
newly-pruned terminal nodes appear at the same length as tips, even if they
were originally internal nodes. Default FALSE.}

\item{tippoint_boxplot}{Alternate parameter for determining whether to color
the tippoints and the boxplots. If TRUE, they will match in colors, and if
FALSE, the boxplots will be filled white with black outline and the
tippoints will be black.}
}
\value{
A ggtree object consisting of subtree induced by data and boxplots
corresponding to specified numeric column of data.
}
\description{
This function takes in a data.table of chemicals with classification data and
additional numeric data, and displays selected numeric data grouped by tip
label on the data-induced subtree of the classification taxonomy.
}
