% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{classify_by_smiles}
\alias{classify_by_smiles}
\title{Classify chemicals from input data.table using SMILES}
\usage{
classify_by_smiles(datatable)
}
\arguments{
\item{datatable}{A data.table that is the output from classify_datatable.}
}
\value{
A data.table object with classification information for each row.
}
\description{
Classify chemicals from input data.table using SMILES
}
