% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{classify_datatable}
\alias{classify_datatable}
\title{Classify chemicals from input data.table using InChIKey}
\usage{
classify_datatable(datatable)
}
\arguments{
\item{datatable}{A data.table with columns 'PREFERRED_NAME', 'CASRN', 'INCHIKEY}
}
\value{
A data.table object with classifications attached to each row of input data.table
}
\description{
Classify chemicals from input data.table using InChIKey
}
