% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{classify_datatable}
\alias{classify_datatable}
\title{Classify data.table}
\usage{
classify_datatable(data)
}
\arguments{
\item{data}{A data.table with rows corresponding to chemicals. A column named
\code{INCHIKEY} containing the InChIKey for each chemical is required.}
}
\value{
A data.table containing the original data as well as the ClassyFire
classification data for each chemical, when available.
}
\description{
Classify data.table
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Chemical information for Bisphenol A
bpa <- data.table(PREFERRED_NAME = 'Bisphenol A',
                  CASRN = '80-05-7',
                  INCHIKEY = 'IISBACLAFKSPIT-UHFFFAOYSA-N')
bpa_classified <- classify_datatable(bpa)
bpa_classified
\dontshow{\}) # examplesIf}
}
