% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{classify_inchikeys}
\alias{classify_inchikeys}
\title{InChIKey classification}
\usage{
classify_inchikeys(
  inchikeys,
  tax_level_labels = chemont_tax_levels,
  wait_sec = 5
)
}
\arguments{
\item{inchikeys}{Character: A vector of InCHiKeys to be classified}

\item{tax_level_labels}{By default, the list of taxonomy levels for
ClassyFire: \code{kingdom, superclass, class, subclass, level5, ...
 level11}.}

\item{wait_sec}{Numeric: A parameter controlling how many seconds between
qqueries sent to the ClassyFire API server. Default 5, to respect the limit
of 12 requests per second. Setting it any lower than 5 will result in failed
queries.}
}
\value{
A \code{data.frame}; see Details.
}
\description{
This function uses the ClassyFire API to classify chemicals from
an input data.table using the InChIKey chemical identifier.
}
\details{
This function queries ClassyFire's lookup table of pre-classified
InChiKeys to get classifications, using the ClassyFire API.

\itemize{
\item{identifier: The input InCHiKey that was queried. For example,
"XGQJGMGAMHFMAO-UHFFFAOYSA-N"}
\item{smiles: The corresponding SMILES returned by ClassyFire, if any}
\item{inchikey: The InCHiKey returned by ClassyFire, if any. Will be of
format "InCHiKey=XGQJGMGAMHFMAO-UHFFFAOYSA-N"}
\item{classification_version: The version number returned by ClassyFire.
For example, "2.1"}
\item{level: The names of all levels in this InCHiKey's classification.
Will be elements of \code{tax_level_labels}.}
\item{name: The name or label of each level in this InCHiKey's
classification.}
\item{report: A text string reporting the status of the classification,
"ClassyFire returned a classification" if successful; otherwise the report
returned by ClassyFire, or a report about an internal error.}
}

Note that the data.frame is in "long" format, with multiple rows for each
InCHiKey. There is one row for each level of classification for each
InCHiKey. For example, the classification for InCHiKey
"XGQJGMGAMHFMAO-UHFFFAOYSA-N" terminates at level 5, so there will be five
rows for that InCHiKey. The classification for InCHiKey
"PZNXLZZWWBSQQK-UHFFFAOYSA-N" terminates at level 4 (subclass), so there
will be four rows for that InCHiKey.

For use with \code{treecompareR} functions that expect a \code{data.frame} of
classified entities, this \code{data.frame} will need to be reshaped into wider
format, with one row for each InCHiKey and one column for each level of
classification. This can be done, e.g., using \code{tidyr::pivot_wider(dat,
 names_from = "level", values_from = "name")} (where \code{dat} is the
returned \code{data.frame}.) However, be on the lookout for pathological cases
where an InCHiKey is listed with two different labels at the same level.
These occur rarely, but they do occur. \code{tidyr::pivot_wider()} will
throw a warning if this happens -- pay attention to it!
}
\examples{
#an example with good, bad, NA, and blank inputs
classify_inchikeys(
c("SEMRCUIXRUXGJX-UHFFFAOYSA-N",
"PLDWAJLZAAHOGG-UHFFFAOYSA-N",
"BAD-INCHIKEY-X",
NA_character_,
"    ")
)

}
\references{
\insertRef{djoumbou2016classyfire}{treecompareR}
}
\seealso{
\code{\link{classify_structures}}
}
