% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{classify_inchikeys}
\alias{classify_inchikeys}
\title{InChIKey classification}
\usage{
classify_inchikeys(
  inchikeys,
  tax_level_labels = c("kingdom", "superclass", "class", "subclass", "level5",
    "level6", "level7", "level8", "level9", "level10", "level11")
)
}
\arguments{
\item{inchikeys}{A vector of InCHiKeys to be classified}
}
\value{
A data.frame with the same number of rows as the length of input
argument \code{inchikeys}, and columns consisting of "INCHIKEY" (containing input
argument \code{inchikeys}) and one column for each of the ClassyFire taxonomy levels
given in the input argument \code{tax_level_labels}
}
\description{
This function uses the ClassyFire API to classify chemicals from an input
data.table using the InChIKey chemical identifier.
}
\seealso{
\code{\link{classify_by_smiles}}
}
