% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{classify_structures}
\alias{classify_structures}
\title{Query ClassyFire by structure}
\usage{
classify_structures(
  input = NULL,
  tax_level_labels = chemont_tax_levels,
  sizelim = 1000,
  wait_sec = 5,
  ...
)
}
\arguments{
\item{input}{A character vector of structural identifiers: SMILES strings or
InChi strings. May optionally be named. If so, the names will be returned
as a column named  \code{identifier} in the output data.frame. If not
named, the vector itself will be returned as a column named
\code{identifier} in the output data.frame}

\item{tax_level_labels}{By default, the list of taxonomy levels for
ClassyFire: \code{kingdom, superclass, class, subclass, level5, ...
  level11}.}

\item{...}{Other arguments as for \code{\link{query_structure}}.}
}
\value{
A list of \code{data.frame}s with ClassyFire classification results for each input
structural identifier.

Will contain variables \code{identifier}, \code{structure},
one variable named for each taxonomy level in \code{tax_level_labels} containing
the classification at each level), \code{smiles},  \code{inchikey}, and then several
variables with information about the ClassyFire query: \code{report},
\code{classification_version}, \code{id}, \code{label}, \code{classification_status}, \code{query_url} and
\code{query_status}. If classification could not be completed for one or more
structures, they will be listed in the \code{data.frame} with \code{NA_character_}
for all taxonomy levels. Check variables \code{report} and \code{status} for more
information about what may have gone wrong.
}
\description{
This function takes a vector of structural identifiers (SMILES
strings or InChi strings) and queries the ClassyFire API to get
classifications for each one.
}
\details{
This function returns a list of \code{data.frame}s with the following names.
\itemize{
\item \code{classification}: The main ClassyFire classification for the input structures.
\item \code{alternative_parents}: Alternative parents for the input structures.
\item \code{molecular_framework}: Molecular framework descriptors for the input structures.
\item \code{substituents}: Substituents for the input structures.
\item \code{description}: Descriptions for the input structures.
\item \code{external_descriptors}: External descriptors for the input structures, if any.
\item \code{ancestors}: All ancestors for the input structures.
\item \code{predicted_chebi_terms}: Predicted ChEBI terms for the input structures.
\item \code{predicted_lipidmaps_terms}: Predicted LipidMaps terms for the input structures.
}
}
\seealso{
\code{\link{classify_inchikeys}}
}
