% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{classify_structures}
\alias{classify_structures}
\title{Query ClassyFire by structure}
\usage{
classify_structures(
  input,
  label = "query",
  type = "STRUCTURE",
  retry_times = 100,
  retry_query_times = 100,
  queued_wait = 2,
  processing_wait_per_input = 1,
  tax_level_labels = c("kingdom", "superclass", "class", "subclass", "level5",
    "level6", "level7", "level8", "level9", "level10", "level11")
)
}
\arguments{
\item{input}{A character vector of SMILES strings or InChiKeys. May
optionally be named. If so, names will be returned as "identifier" column
in output. If not named, the structural identifiers themselves will be
returned as "identifier" column in output.}

\item{label}{An optional text label for the query. Default "query".}

\item{type}{String giving ClassyFire query type. Default "STRUCTURE".}

\item{retry_times}{Number of times to retry GET command. Default 100.}

\item{retry_query_times}{Max number of attempts to retry an "In Queue" or
"Processing" query (with wait time in between tries). Default 100.}

\item{queued_wait}{Number of seconds to wait before retrying to retrieve
results for a queued query. Default 2.}

\item{processing_wait_per_input}{Number of seconds to wait per input string
before retrying to retrieve results for a query whose status is
"Processing." Default 1.}

\item{tax_level_labels}{The default list of taxonomy levels for ClassyFire.}
}
\value{
A data frame with ClassyFire classifications for each input
structural identifier. Will contain columns \code{identifier},
\code{smiles}, \code{inchikey}, one column for each taxonomy level (defined
in argument \code{tax_level_labels}), and \code{report}. The final
\code{report} column explains what happened if a classification could not
be obtained.
}
\description{
This function takes a vector of structural identifiers (SMILES strings or
InChiKeys) and queries the ClassyFire API to get classifications for each
one.
}
\seealso{
\code{\link{classify_inchikeys}}
}
