% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{cluster_analysis}
\alias{cluster_analysis}
\title{Heatmap cluster analysis}
\usage{
cluster_analysis(
  htmap,
  row_cluster,
  column_cluster,
  level = 2,
  tree_object,
  tree = NULL
)
}
\arguments{
\item{htmap}{A ComplexHeatmap object with hierarchical clustering.}

\item{row_cluster}{Index for the row cluster.}

\item{column_cluster}{Index for the column cluster.}

\item{level}{Alternate parameter indicating the level of depth for labels.}

\item{tree_object}{A phylo object representing a root tree, the taxonomy
being investigated.}

\item{tree}{Alternate parameter, a phylo object representing a rooted tree,
for restricting the labels.}
}
\value{
A list of labels for the row and column clusters, based on the level
specified.
}
\description{
This is a helper function that returns superclasses or classes for specified
clusters in \code{\link{generate_tree_cluster}}. This is used for providing
visual identification of different clades based on the specified taxonomic
level.
}
