% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{compare_similarity_measures}
\alias{compare_similarity_measures}
\title{Compare similarity measures}
\usage{
compare_similarity_measures(n)
}
\arguments{
\item{n}{Each tree has 2n-1 total nodes and tips.}
}
\value{
A data.frame consisting of the mean self-similarity scores for each
tree and similarity measure.
}
\description{
This functions compares the similarity measures of Jaccard, Resnik, Lin, and
Jiang and Conrath on a variety of trees. The tree types include the
caterpillar, star, and balanced trees. The input parameter \code{n} indicates how
many tips for the caterpillar and balanced tree. The star has 2n tips, and
all trees have 2n-1 total nodes and tips.
}
