% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_comparison_dashboard.R
\name{convert_smiles_to_hex}
\alias{convert_smiles_to_hex}
\title{Convert SMILES to hex}
\usage{
convert_smiles_to_hex(smiles = NULL)
}
\arguments{
\item{smiles}{A string representing a chemical in SMILES format}
}
\value{
A string converted for use in API calls of the
\href{https://hazard.sciencedataexperts.com/#/}{Hazard Comparison
Dashboard}.
}
\description{
This is a helper function for converting SMILES string to API acceptable
format. This is used by \code{\link{get_toxprints}} for making API calls to
\href{https://hazard.sciencedataexperts.com/#/}{Hazard Comparison Dashboard}
to retrieve relevant chemical information.
}
