% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{data_set_subtrees}
\alias{data_set_subtrees}
\title{Compare data set subtrees}
\usage{
data_set_subtrees(
  data_1,
  data_2,
  name_1 = "data_1",
  name_2 = "data_2",
  tax_level_labels = chemont_tax_levels,
  tree = NULL,
  show_tips = TRUE,
  adjust_branch_length = TRUE
)
}
\arguments{
\item{data_1}{A data.table of chemicals and their classifications.}

\item{data_2}{A data.table of chemicals and their classifications.}

\item{name_1}{An alternate parameter giving the name of the first data set.}

\item{name_2}{An alternate parameter giving the name of the second data set.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}

\item{tree}{An alternate parameter giving a taxonomy if not using ChemOnt.}

\item{show_tips}{An alternate parameter determining whether to show tip
labels.}

\item{adjust_branch_length}{An alternate parameter determining whether to
resize branches of subtree.}
}
\value{
A list of two ggtree objects.
}
\description{
This function takes in two data.tables of chemicals with classification data,
and creates two plots each of which show a visual comparison of the subtrees
induced by two data sets. The visual comparison consists of branches of one
data-induced subtree colored and labeled based on whether they are
represented by chemicals from the other data set.
}
