% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{df_check}
\alias{df_check}
\title{Parsed data frame check}
\usage{
df_check(this_out, item, entities_identifier)
}
\arguments{
\item{this_out}{A \code{data.frame} consisting of parsed output in one element of the ClassyFire JSON \code{entities} element}

\item{item}{Character: the name of the element of the ClassyFire JSON \code{entities} element}

\item{entities_identifier}{Character vector: all identifiers in \code{entities$identifier}}
}
\value{
A \code{data.frame} either the same as the original \code{this_out}, or renamed, or with NAs filled in if \code{this_out} was empty
}
\description{
Helper function used by \code{\link[=get_item]{get_item()}}
}
\details{
This is a helper function used by \code{\link[=get_item]{get_item()}} and generally should not be called directly by the user.
}
\examples{
# There is no example.

}
\author{
Caroline Ring
}
