% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{display_subtree}
\alias{display_subtree}
\title{Display subtree}
\usage{
display_subtree(
  data_1,
  data_2 = NULL,
  name_1 = NULL,
  name_2 = NULL,
  tree = NULL,
  tax_level_labels = NULL
)
}
\arguments{
\item{data_1}{A data.table consisting of classification data for chemicals.}

\item{data_2}{An alternate parameter giving a second data.table with
classification data for chemicals.}

\item{name_1}{An alternate parameter giving the name of \code{data.table_1}.}

\item{name_2}{An alternate parameter giving the name of \code{data.table_1}.}

\item{tree}{An alternate parameter giving a taxonomy if not using ChemOnt.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}
}
\value{
A ggtree object showing the subtree induced by \code{data_1} (and \code{data_2}
if supplied, and their intersection.).
}
\description{
This function takes in one or two data.tables consisting of chemicals with
classification data and returns a tree diagram indiacting the subtree(s)
induced by the data.table(s).
}
