% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{drop_tips_nodes}
\alias{drop_tips_nodes}
\title{Drop tips and nodes}
\usage{
drop_tips_nodes(
  tree,
  data = NULL,
  labels = NULL,
  nodes = NULL,
  level = NULL,
  keep_descendants = NULL,
  tax_level_labels = chemont_tax_levels
)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class).}

\item{data}{A \code{data.frame} of classified entities. Default \code{NULL}.}

\item{labels}{Character: a set of labels of the subtree. Default \code{NULL}.}

\item{nodes}{Integer: a set of nodes of the subtree. Default \code{NULL}.}

\item{level}{Integer: the level to which the tree should be pruned. The root
is level zero and each subsequent generation of children nodes is one level
greater. Default \code{NULL}.}

\item{keep_descendants}{TRUE/FALSE: whether to keep all descendants of the
input labels/nodes/level. Default \code{NULL}, which means behavior depends on
whether \code{data}, \code{labels}, \code{nodes}, or \code{level} was provided. See
\code{\link[=prune_tree]{prune_tree()}} for more details.}

\item{tax_level_labels}{Levels of the taxonomy in order. Default
\code{\link[=chemont_tax_levels]{chemont_tax_levels()}}.}
}
\value{
A \code{phylo}-class object representing the induced subtree of the data.
}
\description{
Helper function used by \code{\link[=prune_tree]{prune_tree()}}.
}
\details{
This function takes in a data set, collects the labels from classifications
of the chemicals in the data set, and drops tips and nodes until the induced
subtree that remains consists solely of the labels associated to the data set
and their ancestors. Alternatively, one can provide a set of labels instead
and the induced subtree is constructed from these labels in the same manner.
}
\examples{
drop_tips_nodes(tree = chemont_tree,
data = biosolids_class[1:20,])


}
\references{
\insertRef{apepackage}{treecompareR}
}
