% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{drop_tips_nodes}
\alias{drop_tips_nodes}
\title{Drop tips and nodes}
\usage{
drop_tips_nodes(tree, data)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{data}{A data.table of chemicals with classifications.}
}
\value{
A phylo object representing the induced subtree of the data.
}
\description{
This function is an extension of drop.tip from the APE package. It takes in a
data set, collects the labels from classifications of the chemicals in the
data set, and drops tips and nodes until the induced subtree that remains
consists solely of the labels associated to the data set.
}
