% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{drop_tips_nodes}
\alias{drop_tips_nodes}
\title{Drop tips and nodes}
\usage{
drop_tips_nodes(
  tree,
  data = NULL,
  labels = NULL,
  nodes = NULL,
  level = NULL,
  keep_descendants = NULL,
  tax_level_labels = c("kingdom", "superclass", "class", "subclass", "level5", "level6",
    "level7", "level8", "level9", "level10", "level11")
)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{data}{A data.table of chemicals with classifications.}

\item{labels}{An alternate parameter for a set of labels of the subtree.}

\item{tax_level_labels}{An alternate parameter passed to the
\code{\link{get_terminal_labels}} function.}
}
\value{
A phylo object representing the induced subtree of the data.
}
\description{
This function is an extension of drop.tip from the APE package. It takes in a
data set, collects the labels from classifications of the chemicals in the
data set, and drops tips and nodes until the induced subtree that remains
consists solely of the labels associated to the data set and their ancestors.
Alternatively, one can provide a set of labels instead and the induced
subtree is constructed from these labels in the same manner.
}
\references{
\insertRef{apepackage}{treecompareR}
}
