% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{enumerate_partitions}
\alias{enumerate_partitions}
\title{Enumerate partitions}
\usage{
enumerate_partitions(n)
}
\arguments{
\item{n}{An integer for which partitions will be generated and enumerated.}
}
\value{
A name list of partitions, with names corresponding to probabilities.
}
\description{
Enumerate partitions
}
