% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{general_Jaccard_similarity}
\alias{general_Jaccard_similarity}
\title{Jaccard similarity}
\usage{
general_Jaccard_similarity(tree, label_A, label_B)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{label_A}{The first label.}

\item{label_B}{The second label.}
}
\value{
The Jaccard distance of the label sets for the root to node path.
}
\description{
This determines the Jaccard similarity for two input node labels in a given
tree. For each node, there is a set of labels along the unique path from the
root to the label. These sets are compared using Jaccard similarity. For more
information on Jaccard similarity, please consult
\href{https://en.wikipedia.org/wiki/Jaccard_index}{Jaccard Index}.
}
\examples{

tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)

general_Jaccard_similarity(tree = tree, label_A = 't2', label_B = 't4')
general_Jaccard_similarity(tree = tree, label_A = 'n3', label_B = 't8')

}
\seealso{
\code{\link{general_Jaccard_dist}}
}
