% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_balanced}
\alias{generate_balanced}
\title{Generate balanced tree}
\usage{
generate_balanced(n)
}
\arguments{
\item{n}{The number of tips}
}
\value{
A \code{phylo} object representing the generated tree.
}
\description{
This function will generate a rooted binary tree that is balanced (or as close
to balanced) with n tips.
}
