% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_caterpillar}
\alias{generate_caterpillar}
\title{Generate caterpillar tree}
\usage{
generate_caterpillar(n)
}
\arguments{
\item{n}{The number of tips.}
}
\value{
A \code{phylo} object representing the generated tree.
}
\description{
This function generates a caterpillar tree, a rooted binary tree of maximal
depth with n tips.
}
