% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{generate_descendants}
\alias{generate_descendants}
\title{Generates a data.frame of the number of descendants for each node of the input tree}
\usage{
generate_descendants(tree)
}
\arguments{
\item{tree}{A phylo object}
}
\value{
data.frame consisting of the node number, descendants, children, and level
}
\description{
Generates a data.frame of the number of descendants for each node of the input tree
}
