% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{generate_heatmap}
\alias{generate_heatmap}
\title{Generate a heatmap from input matrix and indices.}
\usage{
generate_heatmap(
  tree_object,
  matrix,
  row_indices = NA,
  column_indices = NA,
  row_data,
  column_data,
  name = "Name",
  row_split = 1,
  column_split = 1,
  row_title = "Row title",
  column_title = "Column title"
)
}
\arguments{
\item{tree_object}{A phylo object representing a rooted tree.}

\item{matrix}{A matrix of similarity measure values derived from parameter 'tree_object'.}

\item{row_indices}{The row indices for the matrix.}

\item{column_indices}{The column indices for the matrix.}

\item{row_data}{A data.table object of chemical classifications.}

\item{column_data}{A data.table object of chemical classifications.}

\item{name}{Name of the heatmap}

\item{row_split}{Number of clusters for rows.}

\item{row_title}{Title for rows.}

\item{col_split}{Number of cluster for columns.}

\item{col_title}{Title for columns.}
}
\value{
A Heatmap object.
}
\description{
Generate a heatmap from input matrix and indices.
}
