% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{generate_heatmap}
\alias{generate_heatmap}
\title{Generate similarity heatmap}
\usage{
generate_heatmap(
  tree_object,
  matrix,
  row_indices = NA,
  column_indices = NA,
  row_data,
  column_data,
  name = "Name",
  row_split = NULL,
  column_split = NULL,
  row_title = "Row title",
  column_title = "Column title",
  log_trans = TRUE
)
}
\arguments{
\item{tree_object}{A phylo object representing a rooted tree.}

\item{matrix}{A matrix of similarity measure values derived from parameter
\code{tree_object}.}

\item{row_indices}{The row indices for the matrix.}

\item{column_indices}{The column indices for the matrix.}

\item{row_data}{A data.table or data.frame object of chemical classifications.}

\item{column_data}{A data.table or data.frame object of chemical classifications.}

\item{name}{Name of the heatmap similarity measure.}

\item{row_split}{Number of clusters for rows.}

\item{column_split}{Number of cluster for columns.}

\item{row_title}{Title for rows.}

\item{column_title}{Title for columns.}

\item{log_trans}{Alternate parameter for log transforming numbers of label
occurrence for labels present in each data set.}
}
\value{
A Heatmap object.
}
\description{
This function generates a heatmap from an input similarity matrix and
indices. If row data and column data are used instead, indices are generated
from their classification labels and the labels of the similarity matrix.
}
\seealso{
\code{\link{generate_tree_cluster}}
}
