% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{generate_information_content}
\alias{generate_information_content}
\title{Generates a data.frame of information content}
\usage{
generate_information_content(tree, log_descendants = TRUE)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}
}
\value{
data.frame consisting of node number, children, descendants, level, and information content for each node
}
\description{
Generates a data.frame of information content
}
