% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_partition}
\alias{generate_partition}
\title{Partition generator}
\usage{
generate_partition(n, seed = NA)
}
\arguments{
\item{n}{positive integer at least 2.}

\item{seed}{A seed to allow for replication of results.}
}
\value{
A vector of integers each at least 2 and with sum equal to the input.
}
\description{
This function generates a partition of an input positive integer. Each
constituent of the partition has value at least 2.
2.
}
\examples{

generate_partition(n = 4, seed = 42)
generate_partition(n = 4, seed = 24)

}
