% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_partition_2}
\alias{generate_partition_2}
\title{Partition generator 2}
\usage{
generate_partition_2(n, max_deg = NULL, min_deg = 0)
}
\arguments{
\item{n}{Positive integer.}

\item{max_deg}{Maximal value an element of the partition can take.}

\item{min_deg}{Minimal value an element of the partition can take.}
}
\value{
Vector of generated partition.
}
\description{
Randomly generate partition
}
\details{
This function generates a partition with optional constraints placed on it.
If no such partition can be generated, the function stops and reports this.
}
\examples{

set.seed(42)
generate_partition_2(n = 7, min_deg = 3, max_deg = 5)

}
\author{
Paul Kruse
}
