% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_partition_2}
\alias{generate_partition_2}
\title{Generates a partition with optional constraints placed on it.}
\usage{
generate_partition_2(n, max_deg = NULL, min_deg = 0)
}
\arguments{
\item{n}{Positive integer.}

\item{max_deg}{Maximal value an element of the partition can take.}

\item{min_deg}{Minimal value an element of the partition can take.}
}
\value{
Vector of generated partition.
}
\description{
Generates a partition with optional constraints placed on it.
}
