% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_partition_3}
\alias{generate_partition_3}
\title{Partition generator 3}
\usage{
generate_partition_3(n, max_deg, min_deg, seed = NA)
}
\arguments{
\item{n}{positive integer.}

\item{max_deg}{Maximum value an element of the partition can take.}

\item{min_deg}{Minimum value an element of the partition can take.}

\item{seed}{A seed to allow for replication of results.}
}
\value{
Vector with partition or error in case specified partition is
impossible.
}
\description{
This function generates a partition of an input positive integer subject to
both minimum degree and maximum degree constraints.
}
