% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_partition_3}
\alias{generate_partition_3}
\title{Generates partition of positive integer, with optional constraints to partition.}
\usage{
generate_partition_3(n, max_deg, min_deg)
}
\arguments{
\item{n}{positive integer.}

\item{max_deg}{Maximum value an element of the partition can take.}

\item{min_deg}{Minimum value an element of the partition can take.}
}
\value{
Vector with partition or error in case specified partition is impossible.
}
\description{
Generates partition of positive integer, with optional constraints to partition.
}
