% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{generate_similarity_matrix}
\alias{generate_similarity_matrix}
\title{Generates a similarity matrix for a given tree and similarity measure}
\usage{
generate_similarity_matrix(tree, similarity = NULL)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{similarity}{A similarity measure function that requires as inputs a tree, label_A, and label_B}
}
\value{
A similarity matrix, which is a symmetric matrix with values in \eqn{[0,1]}.
}
\description{
Generates a similarity matrix for a given tree and similarity measure
}
