% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_star}
\alias{generate_star}
\title{Generate Star tree}
\usage{
generate_star(n)
}
\arguments{
\item{n}{The number of tips.}
}
\value{
A \code{phylo} object representing the generated tree.
}
\description{
This function generates a star tree, a rooted tree of minimal
depth with n tips.
}
