% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_topology}
\alias{generate_topology}
\title{Generate topology}
\usage{
generate_topology(n, rooted = FALSE, max_deg = NULL, min_deg = NULL, seed = NA)
}
\arguments{
\item{n}{The number of tips, a positive integer.}

\item{rooted}{Whether the tree is rooted or not.}

\item{max_deg}{The maximum degree any node can have.}

\item{min_deg}{The minimum degree any node can have (aside from the root, if
rooted).}

\item{seed}{A seed to allow for replication of results.}
}
\value{
A 'phylo' object representing the generated tree.
}
\description{
This function generates a rooted tree with internal nodes of degree at least
two. The sampling of potential topologies and labels assigned to each
topology are not completely random. The root is degree two, and all other
internal nodes have degrees (potentially) governed by the alternate
parameters \code{max_deg} and \code{min_deg}.
}
