% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_tree.R
\name{generate_tree}
\alias{generate_tree}
\title{Generate tree}
\usage{
generate_tree(file_dir, dataframe = NULL)
}
\arguments{
\item{file_dir}{A path to a JSON file}

\item{dataframe}{An optional data.frame containing parent_child relationship}
}
\value{
A list containing a 'phylo' object and data.frame with useful
information for building the phylo object.
}
\description{
This function will convert taxonomy relationships from a data.frame to a
'phylo' object. In addition to the 'phylo' object, there is a data.frame
returned that includes data relevant to the creation of the 'phylo' object.
}
