% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{generate_tree_cluster}
\alias{generate_tree_cluster}
\title{Generate tree visuals highlighting specified row and column cluster from a heatmap.}
\usage{
generate_tree_cluster(
  tree,
  tree_object,
  htmap,
  row_cluster,
  column_cluster,
  isolate_subtree = FALSE,
  show_labels = FALSE,
  show_clades = TRUE,
  highlight_clades = TRUE,
  point_size = 2,
  bar_size = 1
)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{tree_object}{A phylo object representing the entire taxonomy from which the parameter 'tree' is derived.}

\item{htmap}{A ComplexHeatmap object.}

\item{row_cluster}{Index for row cluster of htmap to be illustrated.}

\item{column_cluster}{Index for column cluster of htmap to be illustrated.}

\item{isolate_subtree}{Alternate parameter for pruning tree diagram to create a second diagram.}

\item{show_labels}{Alternate parameter specifying whether to show tip labels.}

\item{show_clades}{Alternate parameter for labeling superclass clades.}

\item{highlight_clades}{Alternate parameter for labeling superclass clades.}

\item{point_size}{Alternate parameter for size of tip points of represented tips.}

\item{bar_size}{Alternate parameter for size of bars highlighting clades.}
}
\value{
A ggtree object or list of ggtree objects.
}
\description{
Generate tree visuals highlighting specified row and column cluster from a heatmap.
}
