% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{generate_tree_cluster}
\alias{generate_tree_cluster}
\title{Generate tree cluster}
\usage{
generate_tree_cluster(
  tree,
  tree_object,
  htmap,
  row_cluster,
  column_cluster,
  row_name = "Row data set",
  column_name = "Column data set",
  isolate_subtree = FALSE,
  show_labels = FALSE,
  show_clades = TRUE,
  highlight_clades = TRUE,
  point_size = 2,
  bar_size = 1
)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{tree_object}{A phylo object representing the entire taxonomy from which
the parameter \code{tree} is derived.}

\item{htmap}{A ComplexHeatmap object.}

\item{row_cluster}{Index for row cluster of htmap to be illustrated.}

\item{column_cluster}{Index for column cluster of htmap to be illustrated.}

\item{row_name}{Alternate parameter for name of row data set.}

\item{column_name}{Alternte parameter for name of column data set.}

\item{isolate_subtree}{Alternate parameter for pruning tree diagram to create
a second diagram.}

\item{show_labels}{Alternate parameter specifying whether to show tip labels.}

\item{show_clades}{Alternate parameter for labeling superclass clades.}

\item{highlight_clades}{Alternate parameter for labeling superclass clades.}

\item{point_size}{Alternate parameter for size of tip points of represented
tips.}

\item{bar_size}{Alternate parameter for size of bars highlighting clades.}
}
\value{
A ggtree object or list of ggtree objects.
}
\description{
This function generates tree visuals highlighting specified row and column
clusters produced from \code{\link{generate_heatmap}}. The \code{tree} parameter
gives an underlying tree that will be used in the plots. A second plot may
also be produced that prunes away extraneous superclasses to highlight only
the portions of the tree with labels present.
}
\seealso{
\code{\link{generate_heatmap}}
}
