% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Ancestors}
\usage{
get_ancestors(tree, label, node_number = NULL)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{label}{The node label.}

\item{node_number}{Alternate parameter, the number of the given node.}
}
\value{
A list of nodes back to the root of ancestors for the given node.
}
\description{
Generates a list of ancestors for a given node in a tree.
}
\examples{

tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)

get_ancestors(tree = tree, label = 't2')
get_ancestors(tree = tree, label = 'n1')
}
