% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_clade}
\alias{get_clade}
\title{Get node number of clade (ancestor at a specified level)}
\usage{
get_clade(node, tree, level)
}
\arguments{
\item{node}{The node number(s) for which to get the clade(s)}

\item{tree}{The underlying tree (node numbers refer to this tree)}

\item{level}{The hierarchical taxonomy level at which to get the clade(s). Root
is level 0. Default value is 2 (superclass level, in ChemOnt).}
}
\description{
Helper function to get the node number defining a clade for a specified input
node number
}
