% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_cutoffs}
\alias{get_cutoffs}
\title{Similarity cutoffs}
\usage{
get_cutoffs(
  mat,
  data,
  tax_level_labels = chemont_tax_levels,
  neighbors = 3,
  cutoff = NA_real_,
  labels = NULL,
  counts = NULL
)
}
\arguments{
\item{mat}{A similarity matrix corresponding to a similarity measure and a
rooted tree.}

\item{data}{A data.frame of classified entities.}

\item{tax_level_labels}{Parameter giving classification levels.}

\item{neighbors}{A parameter giving how many neighbors to use for finding
label average values.}

\item{cutoff}{Numeric: the cutoff percentage value.}

\item{labels}{Character: a list of node labels
corresponding to a subtree of a rooted tree.}

\item{counts}{Integer: the counts of occurrence for each
label.}
}
\value{
Named list of percentage of data represented by similarity values.
The names are the similarity values. The values of the list are percentages
of data represented by allowing similarity values equal to the names.
}
\description{
This function gives cutoffs for similarity values and percent of
representation of a data set. The function determines what percentage of a
data set that is represented by the induced subtree of the data set for
various values of a fixed similarity measure.
}
\examples{
get_cutoffs(mat = chemont_jaccard, data = biosolids_class)
get_cutoffs(mat = chemont_jaccard, data = biosolids_class, neighbors = 6)

}
