% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_tree.R
\name{get_label_from_node}
\alias{get_label_from_node}
\title{Get label for a tip or internal node of a phylo tree}
\usage{
get_label_from_node(node, tree)
}
\arguments{
\item{node}{Vector of node numbers in phylo tree}

\item{tree}{phylo tree object}
}
\value{
Character vector of tip or internal node labels
}
\description{
Get label for a tip or internal node of a phylo tree
}
