% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_label_from_node}
\alias{get_label_from_node}
\title{Get label for a tip or internal node ID in a phylo tree}
\usage{
get_label_from_node(node, tree)
}
\arguments{
\item{node}{Vector of node ID numbers in phylo tree.}

\item{tree}{phylo tree object}
}
\value{
Character vector of tip or internal node labels corresponding to each
node ID. \code{NA_character_} if no node label corresponds to the input
node ID.
}
\description{
Get label for a tip or internal node ID in a phylo tree
}
