% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{get_label_length}
\alias{get_label_length}
\title{Label length}
\usage{
get_label_length(label_list)
}
\arguments{
\item{label_list}{A named list of labels corresponding to taxonomy levels.}
}
\value{
The number of labels per taxonomy level.
}
\description{
This is a helper function that returns the number of labels per taxonomy
level for a given data.table of chemicals and their classification data.
}
