% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{get_label_level}
\alias{get_label_level}
\title{Label level}
\usage{
get_label_level(data, level_label, tax_level_labels = NULL)
}
\arguments{
\item{data}{A data.table with data that has been classified by some taxonomy.}

\item{level_label}{A string indicating a taxonomy level of the classified
data.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}
}
\value{
The unique labels corresponding to the given level and classified
data.
}
\description{
This function takes in a data.table of chemicals with classification data and
a taxonomy level, and returns all the unique labels for the given taxonomy
level.
}
