% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{get_labels}
\alias{get_labels}
\title{Get labels}
\usage{
get_labels(data, tax_level_labels = chemont_tax_levels)
}
\arguments{
\item{data}{A data.frame consisting of classified items. Rows are items; columns must include all names in \code{tax_level_labels}.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}
}
\value{
A list of terminal classification labels for each item in the input data.
}
\description{
This is a helper function for retrieving labels in a data.frame of classified
chemicals.
}
