% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{get_labels}
\alias{get_labels}
\title{Get labels}
\usage{
get_labels(data, tax_level_labels = NULL)
}
\arguments{
\item{data}{A data.table consisting of classified chemicals.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}
}
\value{
A list of classification labels for each level of taxonomy.
}
\description{
This is a helper function for retrieving labels in a data.table of classified
chemicals, grouped by taxonomy level.
}
