% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_levels}
\alias{get_levels}
\title{Get levels}
\usage{
get_levels(tree)
}
\arguments{
\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class)}
}
\value{
A \code{data.frame} with variables \code{node} (the node number) and \code{level}
(the level of each node in the tree). There is one row for each node in
\code{tree} (including both tips and internal nodes).
}
\description{
Get levels for each node in a tree
}
\details{
Levels are counted starting with the root of the tree at level 0. The
immediate children of the root node are level 1. Then their children are
level 2, etc.
}
\examples{

  tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)

  get_levels(tree = tree)

}
