% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_levels}
\alias{get_levels}
\title{Generate tree levels}
\usage{
get_levels(tree)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}
}
\value{
data.frame consisting of the node number, and the level of each node
}
\description{
This function generates a data.frame of node levels for input rooted tree.
The data.frame has a row corresponding to each tip and internal node. It
records the depth from the root of each node.
}
\examples{

tree <- generate_topology(n = 8, rooted = TRUE, seed = 42)

get_levels(tree = tree)

}
