% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_levels}
\alias{get_levels}
\title{Generates a data.frame of node levels for input rooted tree.}
\usage{
get_levels(tree)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}
}
\value{
data.frame consisting of the node number, and the level of each node
}
\description{
Generates a data.frame of node levels for input rooted tree.
}
