% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_node_from_label}
\alias{get_node_from_label}
\title{Get node ID for a label in a phylo tree}
\usage{
get_node_from_label(label, tree)
}
\arguments{
\item{label}{Vector of labels for tips or internal nodes in phylo tree}

\item{tree}{phylo tree object}
}
\value{
Numeric vector of tip or internal node ID numbers
}
\description{
Get node ID for a label in a phylo tree
}
