% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{get_number_of_labels}
\alias{get_number_of_labels}
\title{Helper function to determine number of occurrences for each label}
\usage{
get_number_of_labels(data, tax_level_labels = NULL)
}
\arguments{
\item{data}{A data.table of chemicals with classifications.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not  using ClassyFire taxonomy}
}
\value{
A named list of chemical labels and their number of occurrences.
}
\description{
Helper function to determine number of occurrences for each label
}
