% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_tree.R
\name{get_path_to_root}
\alias{get_path_to_root}
\title{Helper function for creating tree}
\usage{
get_path_to_root(df, x)
}
\arguments{
\item{df}{data.frame with parent-child relationships}

\item{x}{Numeric identification number for node}
}
\value{
Path from node to root, as a string delimited by '/'.
}
\description{
Helper function for creating tree
}
