% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{get_subtree_nodes}
\alias{get_subtree_nodes}
\title{Get subtree node numbers}
\usage{
get_subtree_nodes(
  data,
  base_tree = chemont_tree,
  tax_level_labels = chemont_tax_levels
)
}
\arguments{
\item{data}{A classified data set.}

\item{tax_level_labels}{A vector of levels for the taxonomy.}

\item{tree}{A taxonomy tree object, as returned by
\code{\link{generate_taxonomy_tree}}.}
}
\value{
A vector of node numbers in the base tree that are represented in the
subtree corresponding to the data set.
}
\description{
This is a helper function that takes a classified set and a taxonomy tree and
provides all node numbers in the subtree corresponding to the data set.
}
