% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_subtree_nodes}
\alias{get_subtree_nodes}
\title{Get subtree node numbers}
\usage{
get_subtree_nodes(
  data,
  base_tree = chemont_tree,
  tax_level_labels = chemont_tax_levels
)
}
\arguments{
\item{data}{A classified data set.}

\item{base_tree}{A phylo-class tree object representing the base tree. Default is \code{\link{chemont_tree}},
the full ChemOnt taxonomy tree.}

\item{tax_level_labels}{A vector of levels for the taxonomy. Default is
\code{\link{chemont_tax_levels}}, the levels of the ChemOnt taxonomy.}
}
\value{
A vector of node numbers in the base tree that are represented in the
subtree corresponding to the data set.
}
\description{
This is a helper function that takes a classified set and a base taxonomy tree and
provides all node numbers in the subtree corresponding to the data set.
}
