% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{get_terminal_labels}
\alias{get_terminal_labels}
\title{Get terminal labels}
\usage{
get_terminal_labels(
  data,
  entity_id_cols = NULL,
  tax_level_labels = chemont_tax_levels
)
}
\arguments{
\item{data}{A \code{data.frame} of classified entities. Variables must include at
least one of the taxonomy level names in \code{tax_level_labels}.}

\item{entity_id_cols}{Character: the name of the variable in \code{data} with
unique entity identifiers. If \code{NULL} (default), then each row is assumed to
be a unique entity.}

\item{tax_level_labels}{A vector of taxonomy levels. Default is
\code{\link{chemont_tax_levels}}, the levels of the ClassyFire taxonomy.}
}
\value{
A vector of terminal classification labels, one for each entity in
the input data.
}
\description{
Retrieve terminal labels in a \code{data.frame} of classified entities.
}
\examples{
#get terminal labels for the first ten chemicals in the classified biosolids set
get_terminal_labels(data = biosolids_class[1:10, ])

}
