% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_methods.R
\name{get_terminal_labels}
\alias{get_terminal_labels}
\title{Get terminal labels}
\usage{
get_terminal_labels(
  data,
  entity_id_cols = NULL,
  tax_level_labels = chemont_tax_levels
)
}
\arguments{
\item{data}{A data.frame consisting of classified items. Rows are entities;
columns must include all names in \code{tax_level_labels}.}

\item{entity_id_cols}{An alternate parameter giving a column name specifying
the id's of the analytes in each row.}

\item{tax_level_labels}{A vector of taxonomy levels. Default is
\code{\link{chemont_tax_levels}}, the levels of the ClassyFire taxonomy.}
}
\value{
A vector of terminal classification labels, one for each entity in
the input data.
}
\description{
This is a helper function for retrieving terminal labels in a data.frame of
classified chemicals.
}
