% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_tip_level}
\alias{get_tip_level}
\title{Returns the tree level of the given node}
\usage{
get_tip_level(tree, label, node_number = NULL)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{label}{The node label.}

\item{node_number}{Alternate parameter, the number of the given node.}
}
\value{
The level of the node from the root of the tree.
}
\description{
Returns the tree level of the given node
}
