% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_comparison_dashboard.R
\name{get_toxprints}
\alias{get_toxprints}
\title{Get toxprints}
\usage{
get_toxprints(smiles)
}
\arguments{
\item{smiles}{A string representing a chemical in SMILES format}
}
\value{
A list of chemical descriptors, including ToxPrints descriptors
}
\description{
This is a function for retrieving toxprint data for a given chemical from the
\href{https://hazard.sciencedataexperts.com/#/}{Hazard Comparison Dashboard}
API to retrieve relevant chemical information.
}
