% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{get_tree_df}
\alias{get_tree_df}
\title{Get data.frame representation of phylo tree}
\usage{
get_tree_df(tree)
}
\arguments{
\item{tree}{A phylo-class tree object}
}
\value{
A data.frame with as many rows as the number of nodes in \code{tree},
and three variables: "node" (the node number), "level" (the hierarchical
level of each node, where the root is level 0), and "Name" (the label
corresponding to each node number).
}
\description{
Helper function to get a data.frame listing node numbers, node names, and
level of each node
}
