% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{handle_missing_node_highlight_clade}
\alias{handle_missing_node_highlight_clade}
\title{Helper function to display clade highlight associated with missing node.}
\usage{
handle_missing_node_highlight_clade(
  tree,
  tree_object,
  list_superclasses,
  tree_visual,
  i,
  color
)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{tree_object}{A phylo object representing the entire taxonomy from which parameter 'tree' is derived.}

\item{list_superclasses}{A list of superclasses to be highlighted.}

\item{tree_visual}{A ggtree object that will have clades highlighted.}

\item{i}{The index for which list_superclasses element will be highlighted.}

\item{color}{A color for the clade highlight.}
}
\value{
A ggtree object that will have the specified clade highlighted.
}
\description{
Helper function to display clade highlight associated with missing node.
}
