% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{handle_missing_node_show_clade}
\alias{handle_missing_node_show_clade}
\title{Helper function to display clade label associated with missing node.}
\usage{
handle_missing_node_show_clade(tree, list_superclasses, tree_visual, i, color)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{list_superclasses}{A list of superclasses to be labeled.}

\item{tree_visual}{A ggtree object that will have clades labeled.}

\item{i}{The index for which list_superclasses element will be labeled.}

\item{color}{A color for the clade label.}
}
\value{
A ggtree object that will have the specified clade labeled.
}
\description{
Helper function to display clade label associated with missing node.
}
