% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Jaccard similarity}
\usage{
jaccard_similarity(
  tree = NULL,
  label_1 = NULL,
  label_2 = NULL,
  node_1 = NULL,
  node_2 = NULL
)
}
\arguments{
\item{tree}{A \code{phylo} object representing a rooted tree.}

\item{label_1}{A label for a node within the \code{tree} object.}

\item{label_2}{A label for a node within the \code{tree} object.}

\item{node_1}{An alternate parameter for a node within the \code{tree} object.}

\item{node_2}{An alternate parameter for a node within the \code{tree} object.}
}
\value{
The Jaccard similarity of the label sets for the root to node path.
}
\description{
This function takes in a tree and two input nodes (either labels or node
numbers) and returns the Jaccard similarity values of the nodes based on the
tree structure. For each node, there is a set of labels along the unique path from the
root to the label. These sets are compared using Jaccard similarity. For more
information on Jaccard similarity, please consult
\href{https://en.wikipedia.org/wiki/Jaccard_index}{Jaccard Index}.
}
\details{
Furthermore, this function is a wrapper for a RCPP function. For rapid
calculation of similarity values of several pairs of nodes, consider using
\code{\link{similarity_matrix}}.
}
\references{
\insertRef{pekar2002taxonomy}{treecompareR}

\insertRef{pesquita2009semantic}{treecompareR}
}
\seealso{
\code{\link{resnik_similarity}}, \code{\link{lin_similarity}},
\code{\link{jiang_conrath_similarity}}, \code{\link{similarity_matrix}}
}
