% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{label_bars}
\alias{label_bars}
\title{Label bars}
\usage{
label_bars(data = NULL, tax_level_labels = NULL)
}
\arguments{
\item{data}{A data.table or list of data.tables with chemicals and their
classification data.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}
}
\value{
A list of two plots, the first showing number of labels by taxonomy
level for each data.table and the second showing the number of labels by
data.table for each taxonomy level.
}
\description{
This function takes in a (list of) data.table(s) and returns figures
illustrating the label numbers by taxonomy level.
}
