% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{label_numbers}
\alias{label_numbers}
\title{Label numbers}
\usage{
label_numbers(
  dat,
  entity_id = NULL,
  tax_level_labels = chemont_tax_levels,
  log10_count = TRUE
)
}
\arguments{
\item{dat}{A data.frame object of classified entities, containing one row for
each entity and one column for each level of the taxonomy.}

\item{entity_id}{The name of the variable in \code{dat} that uniquely
identifies entities. If NULL (default), each row is assumed to be a unique
entity.}

\item{tax_level_labels}{A vector of names for the levels of the taxonomy.
Default is to use the ChemOnt taxonomy levels.}

\item{log10_count}{Optional parameter: TRUE to return log10-transformed label counts;
FALSE to return non-transformed label counts. Default TRUE.}
}
\value{
Number of occurrences of each label in the data.
}
\description{
This is a helper function, use for determining number of labels in a given
data set per taxonomical level.
}
