% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{label_numbers}
\alias{label_numbers}
\title{Helper function for determining number of labels in a given data set}
\usage{
label_numbers(datatable, chemont = TRUE, log = TRUE)
}
\arguments{
\item{datatable}{A data.table object of chemical classifications}
}
\value{
Number of occurrences of each label in the parameter 'datatable'.
}
\description{
Helper function for determining number of labels in a given data set
}
