% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps_and_subplots.R
\name{label_numbers}
\alias{label_numbers}
\title{Label numbers}
\usage{
label_numbers(datatable, chemont = TRUE, log = TRUE)
}
\arguments{
\item{datatable}{A data.table object of chemical classifications.}

\item{chemont}{Alternate parameter indicating whether ChemOnt taxonomy is used.}

\item{log}{Alternate parameter indicating whether numbers are reported as is or as their log.}
}
\value{
Number of occurrences of each label in the parameter \code{datatable}.
}
\description{
This is a helper function, use for determining number of labels in a given
data set per taxonomical level.
}
