% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{lin_similarity}
\alias{lin_similarity}
\title{Lin Similarity}
\usage{
lin_similarity(
  tree = NULL,
  label_1 = NULL,
  label_2 = NULL,
  node_1 = NULL,
  node_2 = NULL
)
}
\arguments{
\item{tree}{The underlying tree being examined.}

\item{label_1}{Set of node labels for rows.}

\item{label_2}{Set of node labels for columns.}

\item{node_1}{Alternate parameter for set of node numbers for rows.}

\item{node_2}{Alternate parameter for set of node numbers for columns.}
}
\value{
The Lin similarity value between the two input nodes.
}
\description{
This function takes in a tree and two input nodes (either labels or node
numbers) and returns the Lin similarity values of the nodes
based on the tree structure. The function uses the formulation as described
in
\href{https://www.researchgate.net/publication/220837848_An_Intrinsic_Information_Content_Metric_for_Semantic_Similarity_in_WordNet/stats}{An
Intrinsic Information Content Metric for Semantic Similarity in WordNet}.
Furthermore, this function is a wrapper for a RCPP function. For rapid
calculation of similarity values of several pairs of nodes, consider using \link{similarity_matrix}.
}
\references{
\insertRef{lin1998information}{treecompareR}
}
\seealso{
\link{resnik_similarity, jiang_conrath_similarity, similarity_matrix}
}
