% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{MonteCarlo_similarity}
\alias{MonteCarlo_similarity}
\title{Random subtree similarity}
\usage{
MonteCarlo_similarity(
  n1,
  n2,
  fun = mean,
  tree = chemont_tree,
  sim_tree = chemont_jaccard
)
}
\arguments{
\item{n1}{Number of nodes to sample for tree1.}

\item{n2}{Number of nodes to sample for tree2.}

\item{tree}{An object of class \code{phylo} representing a rooted tree (see
\code{\link[ape:read.tree]{ape::read.tree()}} for description of this class). Default \code{chemont_tree}
to use the full ChemOnt tree.}

\item{sim_tree}{A pre-computed similarity matrix for \code{tree}. Default
\link{chemont_jaccard} to use the full Jaccard similarity matrix for
\link{chemont_tree}. Other pre-computed options for \link{chemont_tree} include
\link{chemont_resnik_IC_SVH}, \link{chemont_lin_IC_SVH},
\link{chemont_jiangconrath_IC_SVH}.}
}
\value{
A data.frame with similarity values for each simulation. The
similarity values reported in each row is the mean similarity value for the
corresponding data set/simulated tree given by the column.
}
\description{
Random subtree similarity simulation
}
\details{
Given a tree, randomly sample (with replacement) \code{n1} nodes in the tree and
treat these as terminal labels for a subtree (tree1). Repeat for \code{n2},
generating a second randomly sampled subtree tree2. Using a pre-calculated
matrix of similarities for all nodes in the full tree, subset its rows to
keep only the labels from tree1, and its columns to keep only the labels from
tree2.
}
\examples{
set.seed(42)
#average similarity between two random subtrees
#the same sizes as BIOSOLIDS2021 and USGS_WATER
MonteCarlo_similarity(n1 = nrow(biosolids_class), n2 = nrow(usgs_class))

}
\author{
Paul Kruse, Caroline Ring
}
