% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{MonteCarlo_similarity}
\alias{MonteCarlo_similarity}
\title{This function simulates trees for similarity.}
\usage{
MonteCarlo_similarity(
  tree,
  data_1,
  data_2,
  name_1 = "Data set 1",
  name_2 = "Data set 2",
  label_number = 100,
  repetition = 10,
  seed = NA_real_,
  only_tips = FALSE
)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{data_1}{A data.table of chemicals with classifications.}

\item{data_2}{A data.table of chemicals with classifications.}

\item{name_1}{An alternate parameter for the name of \code{data_1}.}

\item{name_2}{An alternate parameter for the name of \code{data_2}.}

\item{label_number}{Number of labels to use to build the simulated trees.}

\item{repetition}{Number of simulated trees to build.}

\item{seed}{Alternate parameter to allow for replication of results.}

\item{only_tips}{Alternate parameter restricting starting labels to tips or
to tips and internal nodes.}
}
\value{
A data.frame with similarity values for each simulation.
}
\description{
This function simulates trees for similarity.
}
