% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{MonteCarlo_similarity}
\alias{MonteCarlo_similarity}
\title{Random subtree similarity simulation}
\usage{
MonteCarlo_similarity(
  tree,
  data_1,
  data_2,
  data_1_indices = NULL,
  data_2_indices = NULL,
  name_1 = "data_set_1",
  name_2 = "data_set_2",
  label_number = 100,
  repetition = 10,
  seed = NA_real_,
  only_tips = FALSE,
  Jaccard = NULL,
  Resnik = NULL,
  Lin = NULL,
  JiangConrath = NULL
)
}
\arguments{
\item{tree}{A phylo object representing a rooted tree.}

\item{data_1}{A data.table of chemicals with classifications.}

\item{data_2}{A data.table of chemicals with classifications.}

\item{data_1_indices}{Alternate parameter giving indices of nodes of a
subtree of \code{tree}.}

\item{data_2_indices}{Alternate parameter giving indices of nodes of a
subtree of \code{tree}.}

\item{name_1}{An alternate parameter for the name of \code{data_1}.}

\item{name_2}{An alternate parameter for the name of \code{data_2}.}

\item{label_number}{Number of labels to use to build the simulated trees.}

\item{repetition}{Number of simulated trees to build.}

\item{seed}{Alternate parameter to allow for replication of results.}

\item{only_tips}{Alternate parameter restricting starting labels to tips or
to tips and internal nodes.}

\item{Jaccard}{The Jaccard similarity matrix for \code{tree}.}

\item{Resnik}{The Resnik similarity matrix for \code{tree}.}

\item{Lin}{The Lin similarity matrix for \code{tree}.}

\item{JiangConrath}{The JiangConrath similarity matrix for \code{tree}.}
}
\value{
A data.frame with similarity values for each simulation. The
similarity values reported in each row is the mean similarity value for the
corresponding data set/simulated tree given by the column.
}
\description{
This function takes in a tree, two data sets (or indices representing two
subtrees of the tree) and creates random subtrees, calculates their
similarity values, and returns the results in a data.frame. This function
allows for comparison of the similarity of the input data sets (or subtrees)
with similarity of random subtrees of specified sizes.
}
\examples{
\donttest{dt1 <- classify_datatable(data.table::data.table(chemical_list_biosolids_2022_05_10)[1:10,])
dt1 <- classify_by_smiles(dt1)

dt2 <- classify_datatable(data.table::data.table(chemical_list_USGSWATER_2022_05_17)[1:10,])
dt2 <- classify_by_smiles(dt2)

MonteCarlo_similarity(tree = treecompareR:::chemont_tree, data_1 = dt1, data_2 = dt2,
                     name_1 = 'Biosolids', name_2 = 'USGS', seed = 42, Jaccard = chemont_jaccard,
                     Resnik = chemont_resnik_IC_SVH, Lin = chemont_lin_IC_SVH,
                     JiangConrath = chemont_jiangconrath_IC_SVH)
MonteCarlo_similarity(tree = treecompareR:::chemont_tree, data_1 = dt1, data_2 = dt2, name_1 = 'Biosolids',
                     name_2 = 'USGS', label_number = 200, seed = 42, Jaccard = chemont_jaccard,
                     Resnik = chemont_resnik_IC_SVH, Lin = chemont_lin_IC_SVH,
                     JiangConrath = chemont_jiangconrath_IC_SVH)}

}
