% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{parse_classified_entities}
\alias{parse_classified_entities}
\title{Parse classified entities}
\usage{
parse_classified_entities(entities, tax_level_labels = chemont_tax_levels)
}
\arguments{
\item{entities}{JSON element of classified entities as returned by ClassyFire
API query}

\item{tax_level_labels}{ChemOnt taxonomy level labels (default
\code{\link{chemont_tax_levels}})}
}
\value{
A data.frame consisting of rows corresponding to each classified
entry from the \code{entities} input. A data.frame with one row for each entity identifier, and variables
\code{identifier}, \code{smiles}, \code{inchikey}, \code{kingdom}, \code{superclass}, \code{class},
\code{subclass}, \code{level5}, \code{level6}, ... \code{level11}, \code{classification_version},
and \code{report}.
}
\description{
Parse classified entities
}
