% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_classyfireR.R
\name{parse_classified_entities}
\alias{parse_classified_entities}
\title{Parse classified entities}
\usage{
parse_classified_entities(entities, tax_level_labels = chemont_tax_levels)
}
\arguments{
\item{entities}{JSON element of classified entities as returned by ClassyFire
API query}

\item{tax_level_labels}{ChemOnt taxonomy level labels (default
\code{\link{chemont_tax_levels}})}
}
\description{
Parse classified entities
}
