% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{prune_and_display_subtree}
\alias{prune_and_display_subtree}
\title{This function takes a data.table of classified chemicals and returns a tree with optional tree diagram}
\usage{
prune_and_display_subtree(
  data,
  tax_level_labels = NULL,
  tree = NULL,
  show_tips = TRUE,
  no_plot = FALSE
)
}
\arguments{
\item{data}{A data.table with classification data for chemicals.}

\item{tax_level_labels}{An alternate parameter giving the taxonomy levels if
not using ClassyFire taxonomy.}

\item{tree}{An alternate parameter giving a taxonomy if not using ChemOnt.}

\item{show_tips}{An alternate parameter determining whether tip labels are displayed.}

\item{no_plot}{An alternate parameter for returning only the pruned tree without the tree visual.}
}
\value{
A pruned tree or list consisting of a pruned tree and ggtree diagram of the pruned tree.
}
\description{
This function takes a data.table of classified chemicals and returns a tree with optional tree diagram
}
