% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{prune_and_display_subtree}
\alias{prune_and_display_subtree}
\title{Prune and display subtree}
\usage{
prune_and_display_subtree(
  base_tree = chemont_tree,
  prune_to = NULL,
  prune_name = NULL,
  adjust_branch_length = FALSE,
  tax_level_labels = chemont_tax_levels,
  keep_descendants = NULL,
  ...
)
}
\arguments{
\item{base_tree}{The "base tree" to prune and then plot, as a
\code{\link[ape]{phylo}}-class object.  Default is the full ChemOnt taxonomy
tree, \code{\link{chemont_tree}}.}

\item{prune_to}{As for \code{\link{prune_tree}}: What to \emph{keep} from the base
tree (everything else will be pruned away). May be a \code{data.frame} of
classified data; one or more labels in the tree (tip or internal node
labels); one or more node numbers in the tree (tip or internal nodes); or
the name of a taxonomy level (one of the items in \code{tax_level_labels}).
Default is NULL, which results in no pruning being done (i.e., the base tree
is returned as-is). See Details.}

\item{prune_name}{Name of pruned tree to use as plot title. Default NULL
results in no plot title.}

\item{adjust_branch_length}{Whether to adjust branch length so that all
newly-pruned terminal nodes appear at the same length as tips, even if they
were originally internal nodes. Default FALSE.}

\item{tax_level_labels}{Optional: a vector of the taxonomy level labels to be
used. Default value: \code{\link{chemont_tax_levels}}, i.e., the levels of
the ClassyFire taxonomy: \code{c("kingdom", "superclass", "class",
 "subclass", paste0("level", 5:11))}.}

\item{keep_descendants}{Whether to keep descendants of specified nodes in the
subtree or not. The default value of the parameter is NULL and the pruning
behavior follows that of \code{\link{prune_to}}. If the parameter value is
TRUE, all descendant nodes will be kept rather than pruned away and if the
parameter value is FALSE, only the nodes of the subtree will remain while
all other nodes are pruned.}

\item{...}{
  Arguments passed on to \code{\link[=display_subtree]{display_subtree}}
  \describe{
    \item{\code{base_name}}{Will be used as the plot title. Usually this should name or
describe the base tree. Default NULL will result in no plot title.}
    \item{\code{data_1}}{Optional: Highlight branches of the base tree according to their
membership in this list. Default is \code{NULL}, to do no highlighting. If
not \code{NULL}, must be one of the following options: A \code{data.frame}
consisting of a list of entities with classification data; a vector of node
numbers in the base tree; a vector of node labels in the base tree; or a
subtree of \code{base_tree} as a \code{\link[ape]{phylo}}-class object. See
Details.}
    \item{\code{data_2}}{Optional: Highlight branches of the base tree to compare
membership in this list and \code{data_1}. Default is \code{NULL}, to do no
highlighting. If not \code{NULL}, one of the following options: A
\code{data.frame} consisting of a list of chemicals with classification data; a
list of node numbers in the base tree; a list of node labels in the base
tree; or a subtree of \code{base_tree} as a \code{\link[ape]{phylo}}-class
object. See Details.}
    \item{\code{name_1}}{Optional: A string giving the name of the list in \code{data_1},
for plot labeling. Default is "Set 1".}
    \item{\code{name_2}}{Optional: A string giving the name of the list in \code{data_2}
(if any), for plot labeling. Default is "Set 2".}
    \item{\code{layout}}{\code{\link{ggtree}} layout option. Default "circular."}
    \item{\code{base_opts}}{List of parameters with default values that control}
    \item{\code{subtree_mapping}}{List of color values for displaying subtree.}
    \item{\code{clade_opts}}{List of parameters with default values for illustrating
clades within the tree diagram.}
    \item{\code{clade_level}}{The taxonomy level at which to draw clade labels, if any.
Root is level 0. Default is level 2 (superclass, in ChemOnt taxonomy). Set
to NULL to suppress clade labels altogether.}
  }}
}
\value{
A ggtree object visualizing the pruned base tree. If\code{data_1}
and/or \code{data_2} are supplied, branches will be highlighted to indicate
whether they are present in each set, neither, or both.
}
\description{
This function takes in a base tree, a pruning specification as for
\code{\link{prune_tree}}, and other arguments as for
\code{\link{display_subtree}}. It first prunes the base tree using
\code{\link{prune_tree}}, then calls \code{\link{display_subtree}}
substituting the pruned tree as the base tree for plotting.  It returns a tree
diagram consisting of the pruned base tree; if data sets were provided,
branches of the base tree will be highlighted according to their membership in
the provided data sets.
}
