% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{prune_tree}
\alias{prune_tree}
\title{Prune a tree}
\usage{
prune_tree(
  tree,
  prune_to = NULL,
  keep_descendants = NULL,
  adjust_branch_length = FALSE,
  tax_level_labels = chemont_tax_levels
)
}
\arguments{
\item{tree}{The "base tree" to be pruned.}

\item{prune_to}{What to keep. May be a \code{data.frame} of classified data;
one or more labels in the tree (tip or internal node labels); one or more
node numbers in the tree (tip or internal nodes). See Details. Default is
NULL, which results in no pruning being done.}

\item{adjust_branch_length}{Whether to adjust branch length so that all
newly-pruned terminal nodes appear at the same length as tips, even if they
were originally internal nodes. Default FALSE.}

\item{tax_level_labels}{Vector of the possible taxonomy levels that can appear
as column names in \code{prune_to} if it is a \code{data.frame} of
classified data.}
}
\value{
A \code{\link[ape]{phylo}}-class object representing the pruned tree.
}
\description{
Prune a tree to keep only a subtree specified as a classified data set, a
vector of labels, or a vector of node numbers
}
