% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_taxonomy_tree.R
\name{SetPlotHeight}
\alias{SetPlotHeight}
\title{Plot height}
\usage{
SetPlotHeight(node, rootHeight = 100)
}
\arguments{
\item{node}{A data.tree object.}

\item{rootHeight}{A positive number placeholder for root height, default 100.}
}
\value{
A data.tree object with plotHeight attribute.
}
\description{
This helper function allows for fast tree construction. See
\href{https://github.com/gluc/data.tree/issues/92#issuecomment-299571142}{Slow
conversion between data.tree and Newick format} for background on this helper
function.
}
\references{
\insertRef{datatreepackage}{treecompareR}
}
