% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_diagrams.R
\name{side_by_side_trees}
\alias{side_by_side_trees}
\title{Side by side trees}
\usage{
side_by_side_trees(data_left, data_right)
}
\arguments{
\item{data_left}{The left data.table of chemical classifications.}

\item{data_right}{The right data.table of chemical classifications.}
}
\value{
An \code{aplot} consisting of two outer layer \code{ggtree} objects and three
inner layer \code{ggplot2} objects.
}
\description{
This function takes two data sets and construsts both trees and internal
layers for data visualization. The left and right out layer of the diagram
consists of subtrees corresponding to the left and right data sets. The left
and right inner layers display label statistics for the left and right data
sets. The center displays label statistics shared by the left and right data
sets.
}
