% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_methods.R
\name{similarity_matrix}
\alias{similarity_matrix}
\title{Similarity matrix}
\usage{
similarity_matrix(
  labels_1 = NULL,
  labels_2 = NULL,
  nodes_1 = NULL,
  nodes_2 = NULL,
  tree = NULL,
  sim_metric = NA_integer_,
  upper_tri = TRUE
)
}
\arguments{
\item{labels_1}{Set of node labels for rows.}

\item{labels_2}{Set of node labels for columns.}

\item{nodes_1}{Alternate parameter for set of node numbers for rows.}

\item{nodes_2}{Alternate parameter for set of node numbers for columns.}

\item{tree}{The underlying tree being examined.}

\item{sim_metric}{The similarity measure. 1 for Jaccard, 2 for Resnik, 3 for
Lin, 4 for Jiang and Conrath.}

\item{upper_tri}{A boolean determining whether the whole matrix or upper
triangle is computed.}
}
\value{
A similarity matrix. The dimension names for the rows and columns
correspond to the associated node labels of the tree.
}
\description{
This function takes in a list of node labels for rows and for columns (or
node indices), a tree, and a similarity measure, and returns a matrix with
the calculated similarity values for each pair. Note that the default is to
compute the similarity values only for pairs along and above the diagonal, so
when the set of nodes for rows and columns differ, one must turn off this
feature.
}
\seealso{
\code{\link{jaccard_similarity}}, \code{\link{resnik_similarity}},
\code{\link{lin_similarity}}, \code{\link{jiang_conrath_similarity}}
}
