% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{simulate_trees}
\alias{simulate_trees}
\title{Tree simulator}
\usage{
simulate_trees(n_trees = 1, simulation_seed = NA, n, ...)
}
\arguments{
\item{n_trees}{The number of trees to generate.}

\item{simulation_seed}{Optional parameter to allow for replication of
simulations.}

\item{n}{The number of tips each tree will have.}

\item{...}{Parameters passed to the generate_topology() function.}
}
\value{
A list of 'phylo' objects, each representing a simulated tree.
}
\description{
This function generates several trees subject to input constraints.
}
