% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{simulate_trees}
\alias{simulate_trees}
\title{Function to generate several trees}
\usage{
simulate_trees(n_trees = 1, simulation_seed = NA, ...)
}
\arguments{
\item{n_trees}{The number of trees to generate.}

\item{simulation_seed}{Optional parameter to allow for replciation of simulations.}

\item{...}{Parameters passed to the generate_topology() function.}
}
\value{
A list of 'phylo' objects, each representing a simulated tree.
}
\description{
Function to generate several trees
}
